% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_scaf_plot_2.R
\name{align_scaf_plot_2}
\alias{align_scaf_plot_2}
\title{Compare Alignments over two Reference scaffolds}
\usage{
align_scaf_plot_2(gr_ob, scaf)
}
\arguments{
\item{gr_ob}{GRanges object containing pairwise alignment}

\item{scaf}{A list containing two character objetcs of the names of the reference scaffolds you wish to compare on the plot}
}
\value{
scaffold alignment plot, comparing two reference scaffolds
}
\description{
The function plots the pairwise alignments from two of the reference scaffolds. This can be used as an aid in super-scaffolding. X-axis is the position on the reference scaffolds, and y-axis is aligned query scaffold.
}
