% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicBreaks.R
\name{as}
\alias{as}
\title{Conversion from \code{\link[CNEr:Axt-class]{CNEr::Axt}} objects}
\description{
Conversion from \code{\link[CNEr:Axt-class]{CNEr::Axt}} objects
}
\note{
By default, in the \emph{Axt} objects produced by the \code{\link[CNEr:readAxt]{CNEr::readAxt}}
function, the coordinates of the query genome are represented with the same
numerical value as in the original file.  In \emph{Axt} files, when the alignment
is on the minus strand, the query genome coordinates have their origin on
the reverse-complement strand.  Therefore, although in the \emph{CNEr} object
they are stored in a \emph{GRanges} object, they do not represent genomic ranges.
In order to do so, there is a \code{\link[CNEr:fixCoordinates]{CNEr::fixCoordinates}} function.
}
\seealso{
Other Bioconductor API functions: 
\code{\link{GBreaks-class}},
\code{\link{getSeq}()},
\code{\link{pairwiseAlignment}()},
\code{\link{range_GBreaks}()},
\code{\link{reverse}()},
\code{\link{subsetByOverlaps_GBreaks}()}
}
\concept{Bioconductor API functions}
