% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_coverage.R
\name{bp_coverage}
\alias{bp_coverage}
\title{Coverage over Breakpoints}
\usage{
bp_coverage(gr_ob, cov_gr, win)
}
\arguments{
\item{gr_ob}{GRanges object containing the pairwise alignment (or alignments of the desired genome)}

\item{cov_gr}{GRanges object of the point coverage of the reference (or desired) genome. Point coverage information should be in the metadata column "name"}

\item{win}{an even number indicating the range overage which the average coverage should be calculated.}
}
\value{
GRanges object of breakpoints and associated point and averaged coverage information
}
\description{
The function returns a GRanges object containing the breakpoints, and metadata colummns of the associated point and averaged coverage over the breakpoint. The averaged coverage is conducted over a window, and the size of which is determined by the user. The coverage is with regards to the reference genome. Make the query genome the subject of the GRanges object to extract this information for the query genome.
}
