% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_heatmap.R
\name{bp_heatmap}
\alias{bp_heatmap}
\title{Breakpoint Associated Heatmaps}
\usage{
bp_heatmap(gr, window, pattern, ...)
}
\arguments{
\item{gr}{\code{GRanges} object containing pairwise alignment}

\item{window}{Range over which heatmap characteristic is plotted.
Breakpoints will be aligned at the center of this.}

\item{pattern}{Character string of desired pattern/characteristic to be
plotted on heatmap.}

\item{...}{Pass other arguments to \code{\link{get_bps}}.}
}
\value{
Returns a \code{\link[heatmaps:Heatmap]{heatmaps::Heatmap}} object of \code{pattern} around centred breakpoints.
}
\description{
This function generates a heatmap of the specified pattern, over breakpoints
aligned at the centre of the plot. The result can be piped into
\code{\link[heatmaps:smoothHeatmap]{heatmaps::smoothHeatmap}} and then \code{\link[heatmaps:plotHeatmapList]{heatmaps::plotHeatmapList}} or
\code{\link[heatmaps:plotHeatmapMeta]{heatmaps::plotHeatmapMeta}}
}
\note{
The \code{GRanges} object is expected to have a \emph{sequence information}
(see \code{\link[Seqinfo:seqinfo]{Seqinfo::seqinfo}}) that allows the retrieval its corresponding \code{BSgenome}
object via the \code{\link[BSgenome:getBSgenome]{BSgenome::getBSgenome}} function.
}
\examples{
# The plot makes no sense, but that is the best example I have at the moment.
exdata_Sac <- system.file("extdata/SacCer3__SacPar.gff3.gz", package = "GenomicBreaks")
gb <- load_genomic_breaks(exdata_Sac, BSgenome.Scerevisiae.UCSC.sacCer3::Scerevisiae)
bp_heatmap(gb, 200, 'GC', dir = "left") |>
  heatmaps::smoothHeatmap() |> heatmaps::plotHeatmapList()

}
\seealso{
Other plot functions:
\code{\link[=bp_pair_analysis]{bp_pair_analysis()}},
\code{\link[=feature_coverage]{feature_coverage()}},
\code{\link[=makeOxfordPlots]{makeOxfordPlots()}},
\code{\link[=plotApairOfChrs]{plotApairOfChrs()}}

Other heatmap functions:
\code{\link[=bp_pair_analysis]{bp_pair_analysis()}},
\code{\link[=feature_coverage]{feature_coverage()}}
}
\author{
Charlotte West

Charles Plessy
}
\concept{heatmap functions}
\concept{plot functions}
