% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{bpGraph}
\alias{bpGraph}
\title{Breakpoint graph}
\usage{
bpGraph(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
A vector of positive integers representing the position of the
breakpoints in the \code{GBreaks} object.
}
\description{
Represent \emph{GenomicBreaks} objects as \emph{breakpoint graphs}.
}
\details{
In a \code{GBreaks} object properly sorted by its \emph{target} ranges, the \emph{query}
ranges …
}
