% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeRegions.R
\name{bridgeRegions}
\alias{bridgeRegions}
\title{Bridge regions}
\usage{
bridgeRegions(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a new \code{GBreaks} object of shorter length.  Its length will
be zero if no bridge regions were found.
}
\description{
Maps unaligned regions of the \emph{target} to the \emph{query} genome when they are
flanked by colinear regions.
}
\note{
Because some aligned regions can be directly adjacent (no gaps), the
returned \code{GBreaks} object may contain ranges of width zero, where the \emph{start}
coordinate is 1 nucleotide higher than the \emph{end} coordinate.
}
\examples{
exampleColinear5
bridgeRegions(exampleColinear5)

# Note the zero-width ranges when aligned regions are directly adjacent.
exampleColinear3
bridgeRegions(exampleColinear3)

}
\references{
Bridge regions have also been called \dQuote{simultaneous gaps}
in the comparison of the mouse and human genomes by Kent WJ, Baertsch R,
Hinrichs A, Miller W, Haussler D. (\emph{Evolution's cauldron: duplication,
deletion, and rearrangement in the mouse and human genomes.} Proc Natl Acad
Sci U S A. 2003;100(20):11484-11489. doi:10.1073/pnas.1932072100)
}
\seealso{
Other Colinearity functions:
\code{\link[=GOC]{GOC()}},
\code{\link[=chain_contigs]{chain_contigs()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=dist2next]{dist2next()}},
\code{\link[=filterColinearRegions]{filterColinearRegions()}},
\code{\link[=flagColinearAlignments]{flagColinearAlignments()}},
\code{\link[=flagPairs]{flagPairs()}}

Other modifier functions:
\code{\link[=chromosome_mapping]{chromosome_mapping()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\author{
Charles Plessy
}
\concept{Colinearity functions}
\concept{modifier functions}
