% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_mapping.R
\name{chromosome_mapping}
\alias{chromosome_mapping}
\title{Chromosome Mapping}
\usage{
chromosome_mapping(gb, min_block_size_perc = 0)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{min_block_size_perc}{A value between 0 and 100 used to filter the aligned blocks based on their size.
Only aligned blocks with sizes greater than the percentile specified by the parameter are considered for mapping.
For example, \code{min_block_size_perc=10} will ignore the 10\% smallest blocks.}
}
\value{
Returns a list where each element corresponds to a mapped pair of chromosomes.
Each element contains the following:
1. \code{ref_chr_name} : the name of the reference chromosome;
2. \code{qry_chr_name} : the name of the query chromosome;
3. \code{gb} : a \emph{GenomicBreaks} object containing only aligned blocks for the corresponding chromosome
pair that are larger than the threshold specified by \code{min_block_size_perc}.
}
\description{
This function computes a mapping between the chromosomes of the reference and query genomes.
The mapping is chosen to maximize the number of aligned blocks in the mapped chromosome pairs.
The function tries to build a \code{1:1} mapping, but this outcome is not guaranteed (see note below).
}
\note{
The function tries to build a \code{1:1} mapping between chromosomes in the reference and query genomes;
however, in exceptional cases, one chromosome may end up mapped to multiple chromosomes.
When the genomes have unequal chromosome counts, the genome with more chromosomes will necessarily have unpaired chromosomes.
}
\examples{
\dontrun{
	chrom_map <- chromosome_mapping(gb, 10)
	cat("Chromosome mapping:\n")
	for (chrom_pair in chrom_map) {
	   cat(paste("- " , chrom_pair$ref_chr_name, " <-> ", chrom_pair$qry_chr_name, " (Nb. aligned blocks: ", length(chrom_pair$gb@ranges@start), ") \n", sep=""))
	}
	}

}
\seealso{
Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\author{
Priscila Biller
}
\concept{modifier functions}
