% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSubstitutionMatrix.R
\docType{data}
\name{exampleSubstitutionMatrix}
\alias{exampleSubstitutionMatrix}
\alias{exampleProbabilityMatrix}
\title{Example Substitution Matrices}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\usage{
exampleSubstitutionMatrix
}
\description{
The probability matrix is loaded  from the file \code{extdata/example.train} distributed in this package.
}
\details{
The count matrix is derived from an alignment of \emph{Oikopleura dioica} cryptic species…
}
\seealso{
Other Alignment statistics:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GCequilibrium]{GCequilibrium()}},
\code{\link[=GCpressure]{GCpressure()}},
\code{\link[=GCproportion]{GCproportion()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=gapProportion]{gapProportion()}},
\code{\link[=logDet_distance]{logDet_distance()}}
}
\concept{Alignment statistics}
\keyword{datasets}
