% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{filterInversions}
\alias{filterInversions}
\title{Extract central blocks in inversions}
\usage{
filterInversions(gb, rename = TRUE, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}}.}

\item{rename}{Replace range names by their numeric order before subsetting.}

\item{detect}{Run again \code{flagInversions()} if \code{TRUE}, else reuse the \code{inv}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
the central part of an inversion triplet have been discarded.  If the object
was missing the \code{inv} metadata column, return the object after discarding all
of its ranges.
}
\description{
Extract central blocks in inversions
}
\examples{
filterInversions(exampleInversion)

}
\seealso{
Other Inversion functions:
\code{\link[=filterDoubleInversions]{filterDoubleInversions()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=leftInversionGaps]{leftInversionGaps()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=showInversions]{showInversions()}}
}
\concept{Inversion functions}
