% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInsersions.R
\name{flagInsersions}
\alias{flagInsersions}
\title{Flag insersions}
\usage{
flagInsersions(gr_ob, tol = Inf)
}
\arguments{
\item{gr_ob}{A \code{GenomicBreaks} object.}

\item{tol}{Tolerance window for the distance between two ranges.}
}
\value{
Returns the \code{GenomicBreaks} object with an extra \code{ins} metadata
column.
}
\description{
Flag ranges that start a triplet where the reference range is colinear and
the query range is made of two blocks that would be colinear if they were
not separated by an block on a different molecule.
}
\details{
Here is a trivial example of an insersion.\preformatted{┌───────────────┬───────────────┬───────────────┐
│ XSR:101-200:+ │ XSR:201-300:+ │ XSR:301-400:+ │  (OKI2018_I69)
└───────────────┴───────────────┴───────────────┘
┌───────────────┬───────────────┬───────────────┐
│  S1:101-200:+ │  S2:801-900:+ │  S1:201-300:+ │  (OSKA2016)
└───────────────┴───────────────┴───────────────┘
}
}
\examples{
ins       <- GenomicRanges::GRanges(c("XSR:101-200:+", "XSR:201-300:+",  "XSR:301-400:+"))
ins$query <- GenomicRanges::GRanges(c( "S1:101-200",   "S2:801-900",     "S1:201-300"   ))
flagInsersions(ins)

}
\seealso{
Other Flagging functions: 
\code{\link{flagColinearAlignments}()},
\code{\link{flagInversions}()}
}
\concept{Flagging functions}
