% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInterruptions.R
\name{flagInterruptions}
\alias{flagInterruptions}
\title{Flag interruptions}
\usage{
flagInterruptions(gb)
}
\arguments{
\item{gb}{A \code{GenomicBreaks} object.}
}
\value{
Returns the \code{GenomicBreaks} object with an extra \code{int} metadata
column.
}
\description{
Flag ranges that start a triplet where the flanking ranges are colinear if
the central pair is removed.
}
\details{
Here is a trivial example of an interruption\preformatted{┌───────────────┬────────────────┬───────────────┐
│ XSR:101-200:+ │ Chr1:201-300:+ │ XSR:301-400:+ │  (OKI2018_I69)
└───────────────┴────────────────┴───────────────┘
┌───────────────┬────────────────┬───────────────┐
│  S1:101-200:+ │  S2:201-300:+  │  S1:301-400:+ │  (OSKA2016)
└───────────────┴────────────────┴───────────────┘
}
}
\note{
Note that THIS FUNCTION RETURNS \code{FALSE} for \emph{inversions} and
\emph{insertions}.
}
\examples{
# On the plus strand
int       <- GenomicRanges::GRanges(c("XSR:101-200:+", "Chr1:201-300:-",  "XSR:301-400:+"))
int$query <- GenomicRanges::GRanges(c( "S1:101-200",   "S2:201-300",      "S1:301-400"))
flagInterruptions(int)

# On the minus strand
int2       <- GenomicRanges::GRanges(c("XSR:301-400:-", "Chr1:201-300:+",  "XSR:101-200:-"))
int2$query <- GenomicRanges::GRanges(c( "S1:301-400",   "S2:201-300",      "S1:101-200"))
flagInterruptions(int2)

# No interruption
int3       <- GenomicRanges::GRanges(c("XSR:101-200:+", "XSR:201-300:+",  "XSR:301-400:+"))
int3$query <- GenomicRanges::GRanges(c( "S1:101-200",   "S1:201-300",      "S1:301-400"))
flagInterruptions(int3)

}
\seealso{
Other Flagging functions: 
\code{\link{flagAll}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagInsersions}()},
\code{\link{flagInversions}()}
}
\concept{Flagging functions}
