% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flagInversions}
\alias{flagInversions}
\title{Flag inversions}
\usage{
flagInversions(gb, tol = Inf)
}
\arguments{
\item{gb}{A \code{GBreaks()} object.}

\item{tol}{Tolerance window for the distance between two ranges.}
}
\value{
Returns the \code{GBreaks} object with an extra \code{inv} metadata column.
}
\description{
Flag ranges that start a triplet that would be colinear if the central pair
were aligned to the opposite strand.
}
\details{
Here is a trivial example of an inversion.

\if{html}{\out{<div class="sourceCode">}}\preformatted{┌──────────────┬──────────────┬──────────────┐
│ chrA:101-200 │ chrA:201-300 │ chrA:301-400 │ (Target genome)
└──────────────┴──────────────┴──────────────┘
       +               -             +         (Alignment direction)
┌──────────────┬──────────────┬──────────────┐
│ chrB:101-200 │ chrB:201-300 │ chrB:301-400 │ (Query genome)
└──────────────┴──────────────┴──────────────┘
}\if{html}{\out{</div>}}
}
\examples{
flagInversions(exampleInversion)
flagInversions(exampleInversion, tol = 19)
plotApairOfChrs(exampleInversion)

flagInversions(exampleInversion |> reverse() |> sort(ignore.strand = TRUE))
plotApairOfChrs(exampleInversion |> reverse())

}
\seealso{
Other Flagging functions:
\code{\link[=flagAll]{flagAll()}},
\code{\link[=flagColinearAlignments]{flagColinearAlignments()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagLongShort]{flagLongShort()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTranslocations]{flagTranslocations()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}}

Other Inversion functions:
\code{\link[=filterDoubleInversions]{filterDoubleInversions()}},
\code{\link[=filterInversions]{filterInversions()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=leftInversionGaps]{leftInversionGaps()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=showInversions]{showInversions()}}

Other Structural variants:
\code{\link{StructuralVariants}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTranslocations]{flagTranslocations()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=plotApairOfChrs]{plotApairOfChrs()}}
}
\concept{Flagging functions}
\concept{Inversion functions}
\concept{Structural variants}
