% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{flagTranslocations}
\alias{flagTranslocations}
\title{Flag translocations}
\usage{
flagTranslocations(gb, tol = Inf, both = TRUE)
}
\arguments{
\item{gb}{A \code{GBreaks()} object.}

\item{tol}{Tolerance window for the distance between two ranges.}

\item{both}{Flag both the \emph{target} and \emph{query} ranges.}
}
\value{
Returns the \code{GBreaks} object with an extra \code{tra} metadata column.
}
\description{
Flag ranges that start a triplet that would be colinear if the central pair
were colinear, and that are not an inversion.
}
\details{
The result of the flagging differs if the \emph{target} and \emph{query} ranges are
swapped.  Therefore, by default this function will search for translocations
on both cases, and flag both ranges.

Here is a trivial example of a translocation:

\if{html}{\out{<div class="sourceCode">}}\preformatted{┌──────────────┬──────────────┬──────────────┐
│ chrA:101-200 │ chrA:201-300 │ chrA:301-400 │ (Target genome)
└──────────────┴──────────────┴──────────────┘
       +               +             +         (Alignment direction)
┌──────────────┬──────────────┬──────────────┐
│ chrB:101-200 │ chrC:201-300 │ chrB:301-400 │ (Query genome)
└──────────────┴──────────────┴──────────────┘
}\if{html}{\out{</div>}}
}
\examples{
flagTranslocations(exampleTranslocation)
flagTranslocations(exampleTranslocation2)
plotApairOfChrs(exampleTranslocation)
flagTranslocations(exampleDeletion)
flagTranslocations(exampleInsertion)
flagTranslocations(exampleInsertion  |> swap(sort = TRUE))
flagTranslocations(exampleInsertion) |> swap(sort = TRUE)
flagTranslocations(sort(reverse(exampleDeletion)))
flagTranslocations(exampleInversion)
flagTranslocations(exampleColinear3)

}
\seealso{
Other Flagging functions:
\code{\link[=flagAll]{flagAll()}},
\code{\link[=flagColinearAlignments]{flagColinearAlignments()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagLongShort]{flagLongShort()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}}

Other Translocation functions:
\code{\link[=filterTranslocations]{filterTranslocations()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=showTranslocations]{showTranslocations()}}

Other Structural variants:
\code{\link{StructuralVariants}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=plotApairOfChrs]{plotApairOfChrs()}}
}
\concept{Flagging functions}
\concept{Structural variants}
\concept{Translocation functions}
