% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flagTwinInversions}
\alias{flagTwinInversions}
\title{Flag twin inversions}
\usage{
flagTwinInversions(gb, details = FALSE)
}
\arguments{
\item{gb}{A \code{GBreaks()} object.}

\item{details}{Returns more metadata columns if \code{TRUE}.}
}
\value{
Returns the \code{GBreaks} object with an extra \code{twi} metadata column.
}
\description{
Flag ranges that start a quadruplet that would be colinear if the two central
pairs were aligned to the opposite strand.  This implies breakpoint reuse.
}
\details{
Here is a trivial example of a twin inversion.

\if{html}{\out{<div class="sourceCode">}}\preformatted{┌──────────────┬──────────────┬──────────────┐──────────────┐
│ chrA:101-200 │ chrA:201-300 │ chrA:301-400 │ chrA:401-500 │ (Target genome)
└──────────────┴──────────────┴──────────────┘──────────────┘
       +               -             -              +   (Alignment direction)
┌──────────────┬──────────────┬──────────────┐──────────────┐
│ chrB:101-200 │ chrB:201-300 │ chrB:301-400 │ chrB:401-500 │  (Query genome)
└──────────────┴──────────────┴──────────────┘──────────────┘
}\if{html}{\out{</div>}}
}
\examples{
flagTwinInversions(exampleTwinInversions)
plotApairOfChrs(exampleTwinInversions)

flagTwinInversions(exampleTwinInversions |> reverse() |> sort(ignore.strand = TRUE))
plotApairOfChrs(exampleTwinInversions |> reverse())

}
\seealso{
Other Flagging functions:
\code{\link[=flagAll]{flagAll()}},
\code{\link[=flagColinearAlignments]{flagColinearAlignments()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagLongShort]{flagLongShort()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTranslocations]{flagTranslocations()}}

Other Inversion functions:
\code{\link[=filterDoubleInversions]{filterDoubleInversions()}},
\code{\link[=filterInversions]{filterInversions()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=leftInversionGaps]{leftInversionGaps()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=showInversions]{showInversions()}}

Other Structural variants:
\code{\link{StructuralVariants}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTranslocations]{flagTranslocations()}},
\code{\link[=plotApairOfChrs]{plotApairOfChrs()}}
}
\concept{Flagging functions}
\concept{Inversion functions}
\concept{Structural variants}
