% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flipInversions}
\alias{flipInversions}
\title{Flip strand of inversions}
\usage{
flipInversions(gb, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{detect}{Run again \code{flagInversions()} if \code{TRUE}, else reuse the \code{inv}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are the
central part of an inversion triplet have their strand orientation flipped.
The \code{inv} metadata column is then discarded as it is not valid anymore.  As
the former inversion triplets are now collinear, new inversions may possibly
found after the object is coalesced again.
}
\description{
Flip strand of inversions
}
\examples{
exampleNestedInversions |> flagInversions()
flipInversions(exampleNestedInversions)

}
\seealso{
Other Inversion functions:
\code{\link[=filterDoubleInversions]{filterDoubleInversions()}},
\code{\link[=filterInversions]{filterInversions()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=leftInversionGaps]{leftInversionGaps()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=showInversions]{showInversions()}}

Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\concept{Inversion functions}
\concept{modifier functions}
