% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandNames.R
\name{flipStrandNames}
\alias{flipStrandNames}
\title{Reverse order and flip strand}
\usage{
flipStrandNames(x)
}
\arguments{
\item{x}{A vector of strand names.}
}
\value{
Returns the same class as its input.
}
\description{
Takes a character vector of "strand names", reverses the order of its
elements and flips the strand.
}
\seealso{
Other scaffolding functions:
\code{\link[=longestMatchesInTarget]{longestMatchesInTarget()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=scaffoldByFlipAndMerge]{scaffoldByFlipAndMerge()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=strandNames]{strandNames()}}
}
\concept{scaffolding functions}
