% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gb2comp}
\alias{gb2comp}
\title{Convert \code{GBreaks} object to a \code{comparison} object}
\usage{
gb2comp(gb, color = NULL, ignore.strand = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{color}{(optional) A vector of colors, of same length as \code{gb}.}

\item{ignore.strand}{Ignore strand information?}
}
\value{
Returns a \code{\link[genoPlotR:comparison]{genoPlotR::comparison}} object.
}
\description{
Convert \code{GBreaks} object to a \code{comparison} object
}
\note{
Note that the \code{comparison} objects assume that the ranges are on the
same sequence feature.  It is your responsibility that the input \code{GBreaks}
object also conforms to this assumption.
}
\examples{
gb2comp(exampleInversion)

}
\seealso{
Other genoPlotR functions:
\code{\link[=gb2dna_seg]{gb2dna_seg()}},
\code{\link[=gb2xlim]{gb2xlim()}},
\code{\link[=gr2dna_seg]{gr2dna_seg()}},
\code{\link[=plotApairOfChrs]{plotApairOfChrs()}}

Other Converter functions:
\code{\link[=GBreaksToMatrix]{GBreaksToMatrix()}},
\code{\link[=gb2dna_seg]{gb2dna_seg()}},
\code{\link[=gb2xlim]{gb2xlim()}},
\code{\link[=gr2dna_seg]{gr2dna_seg()}}
}
\author{
Charles Plessy
}
\concept{Converter functions}
\concept{genoPlotR functions}
