% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gb2xlim}
\alias{gb2xlim}
\title{GBreaks to xlim}
\usage{
gb2xlim(gb)
}
\arguments{
\item{gb}{A \code{GBreaks} object.}
}
\value{
A \code{list} of two pairs of numbers, where the first pair is the \emph{start}
and \emph{end} positions of the first \code{\link{range}} of the \code{GBreaks} object on the
\emph{target} genome and the second pair is the \emph{start} and \emph{end} positions of the
same first range, on the \emph{query} genome.
}
\description{
Converts a \code{GBreaks} object to a \code{list} that can be passed as \code{xlim} argument
to \code{\link{plotApairOfChrs}}.
}
\note{
This function is intended to be run on \code{GBreaks} objects that have been
filtered to contain only one \emph{seqname} on each genome.
}
\examples{
gb2xlim(exampleColinear)

}
\seealso{
Other genoPlotR functions:
\code{\link[=gb2comp]{gb2comp()}},
\code{\link[=gb2dna_seg]{gb2dna_seg()}},
\code{\link[=gr2dna_seg]{gr2dna_seg()}},
\code{\link[=plotApairOfChrs]{plotApairOfChrs()}}

Other Converter functions:
\code{\link[=GBreaksToMatrix]{GBreaksToMatrix()}},
\code{\link[=gb2comp]{gb2comp()}},
\code{\link[=gb2dna_seg]{gb2dna_seg()}},
\code{\link[=gr2dna_seg]{gr2dna_seg()}}
}
\author{
Charles Plessy
}
\concept{Converter functions}
\concept{genoPlotR functions}
