% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBreaksToMatrix.R
\name{GBreaksToMatrix}
\alias{GBreaksToMatrix}
\title{Convert to binned matrix of hits}
\usage{
GBreaksToMatrix(gb, ntile = 1000)
}
\arguments{
\item{gb}{A \code{GenomicBreaks} object.}

\item{ntile}{The number of bins in the matrix.}
}
\value{
A \code{\link{matrix}} object.
}
\description{
Convert to binned matrix of hits
}
\examples{
m <- GBreaksToMatrix(exampleColinear5, ntile = 10)
m
image(m)
makeOxfordPlots(exampleColinear5)

}
\seealso{
Other Converter functions:
\code{\link[=gb2comp]{gb2comp()}},
\code{\link[=gb2dna_seg]{gb2dna_seg()}},
\code{\link[=gb2xlim]{gb2xlim()}},
\code{\link[=gr2dna_seg]{gr2dna_seg()}}
}
\concept{Converter functions}
