% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bps.R
\name{get_bps}
\alias{get_bps}
\title{Get breakpoints}
\usage{
get_bps(
  gr,
  direction = c("both", "left", "right", "mid"),
  stranded = FALSE,
  sorted = TRUE
)
}
\arguments{
\item{gr}{\code{GRanges} object containing pairwise alignment}

\item{direction}{Return the breakpoints on \code{both}, \code{left} or \code{right} side(s)
of the range, or at the \code{mid}point between ranges.}

\item{stranded}{If \code{TRUE}, will assign a \code{+} strand to the left-side
breakpoints and a \code{-} strand to the right-side ones.}

\item{sorted}{Sorts the object before returning it.}
}
\value{
\code{GRanges} object of the breakpoints
}
\description{
Given a \code{\link{GBreaks}} or \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges}} object, the function produces a \code{GRanges}
object detailing the breakpoints only.
}
\examples{
get_bps(exampleInversion)
get_bps(exampleInversion, direction = "left")
get_bps(exampleInversion, stranded = TRUE)
get_bps(exampleInversion, direction = "right", stranded = TRUE)
get_bps(exampleInversion, direction = "mid")

}
\seealso{
Other Reducing functions:
\code{\link[=cleanGaps]{cleanGaps()}},
\code{\link[=wholeGenomeClassification]{wholeGenomeClassification()}}
}
\concept{Reducing functions}
