% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence_time_timetree.R
\name{get_divergence_time}
\alias{get_divergence_time}
\title{Divergence time - TimeTree}
\usage{
get_divergence_time(taxid1, taxid2)
}
\arguments{
\item{taxid1}{An integer: The NCBI id of one taxon}

\item{taxid2}{An integer: The NCBI id of the other taxon}
}
\value{
A list in which the first item is the median divergence time (\code{divtime_med_age}),
and the last two items are the divergence time confidence interval
(\code{divtime_ci_low} and \code{divtime_ci_high}).
If the divergence time is not found in the TimeTree database, it returns \code{NA}.
}
\description{
Retrieves the divergence time of two species via the TimeTree API.
}
\examples{
\dontrun{
get_divergence_time(9555,9601)}

}
\author{
Priscila Biller
}
\concept{Query functions}
