% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergenceTime_Timetree.R
\name{getDivergenceTime}
\alias{getDivergenceTime}
\title{Divergence time - TimeTree}
\usage{
getDivergenceTime(taxid1, taxid2)
}
\arguments{
\item{taxid1}{An integer: The NCBI id of one taxon}

\item{taxid2}{An integer: The NCBI id of the other taxon}
}
\value{
A list in which the first item is the median divergence time (\code{divtimeMedian}),
and the last two items are the divergence time confidence interval
(\code{divtimeCI_low} and \code{divtimeCI_high}).
If the divergence time is not found in the TimeTree database, it returns \code{NA}.
}
\description{
Retrieves the divergence time of two species via the TimeTree API.
}
\examples{
\dontrun{
getDivergenceTime(9555,9601)}

}
\author{
Priscila Biller
}
\concept{Query functions}
