% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefGenome.R
\name{getRefGenome}
\alias{getRefGenome}
\title{Attempt to load a BSgenome}
\usage{
getRefGenome(reference.genome)
}
\arguments{
\item{reference.genome}{name of genome package (and also genome data object)}
}
\value{
A BSgenome object
}
\description{
Attempt to load a BSgenome
}
\details{
Internal function that retrieves a BSgenome object or throws an error if not available. Adapted from the CAGEr package.
}
\author{
Charles Plessy
}
\keyword{internal}
