% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_O_Oki_data.R
\docType{data}
\name{gr_O_Oki}
\alias{gr_O_Oki}
\title{Genome Pairwise Alignment}
\format{GRanges object containing reference genome alignment, with corresponding query alignment information in metadata.}
\usage{
gr_O_Oki
}
\description{
Pairwise alignment between the Osaka2016 O. dioica genome and Okinawa O. dioica genome (v3).
}
\keyword{datasets}
