% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{hurdles_count}
\alias{hurdles_count}
\title{Count hurdles}
\usage{
hurdles_count(g, query_sequence_unsig)
}
\arguments{
\item{g}{The breakpoint graph.}

\item{query_sequence_unsig}{The extended permutation.}
}
\description{
This function finds hurdles in the breakpoint graph and returns them.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=cycle_nontrivial_count]{cycle_nontrivial_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
