% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_frag.R
\name{int_frag}
\alias{int_frag}
\title{Integrated intersection of both query and reference subsets on a continuous alignment}
\usage{
int_frag(og_r, og_q, sub_r, sub_q)
}
\arguments{
\item{og_r}{GRanges object containing the refernce genome information of the original alignment}

\item{og_q}{GRanges object containing the query genome information of the original alignment}

\item{sub_r}{GRanges object containing the reference subset}

\item{sub_q}{GRanges object containing the query subset}
}
\value{
GRanges object of alignment intersected on both reference and query side. Query data held in metadata column
}
\description{
Integrated intersection of both query and reference subsets on a continuous alignment
}
