% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv2UCSCData.R
\name{inv2UCSCData}
\alias{inv2UCSCData}
\title{Represent inversions as UCSCData objects}
\usage{
inv2UCSCData(gb)
}
\arguments{
\item{gb}{A \code{GenomicBreaks} object.}
}
\value{
Returns a \code{UCSCData} object.
}
\description{
Scan a \code{\link{GenomicBreaks}} object for inversions and return UCSC track object
in \code{\link[rtracklayer:UCSCData-class]{rtracklayer::UCSCData}} format for export in BED12 format with
a command such as \code{rtracklayer::export(x, "test.bed", "BED")}.
}
