% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversionEstimate_BD.R
\name{inversionEstimate_BD_single}
\alias{inversionEstimate_BD_single}
\title{Inversion Estimate - option 'single'}
\usage{
inversionEstimate_BD_single(n, obs_nb_cycles)
}
\arguments{
\item{n}{An integer: The number of markers.
A marker represents an aligned block, a gene,
or any other genomic region being rearranged.}

\item{obs_nb_cycles}{An integer: The observed number of cycles
in the breakpoint graph of two genomes.}
}
\description{
It finds the number of inversions whose expected number of cycles
is closest to the observed number of cycles.
}
\details{
This function is an alternative to the function "inversionEstimate_BD_many",
which uses the function "inversionEstimate_BD_setup".

This function should be preferred when there are few inversion estimates to be computed.
}
\author{
Priscila Biller
}
\keyword{internal}
