% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicBreaks.R
\name{isSorted}
\alias{isSorted}
\alias{isSorted,GBreaks-method}
\title{Test if a \code{GBreaks} object is sorted}
\usage{
isSorted(x)

\S4method{isSorted}{GBreaks}(x)
}
\arguments{
\item{x}{A \code{GBreaks} object.}
}
\value{
Returns \code{TRUE} or \code{FALSE}.
}
\description{
The only proper way to sort a \code{\link{GBreaks}} object is by ignoring strand, so
that inversions and deletions are easily detected and visualised.
}
\examples{
isSorted(exampleInversion)

}
