% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JC69_distance_allseq.R
\name{JC69_distance_allseq}
\alias{JC69_distance_allseq}
\title{Variation of the Jukes-Cantor 1969 distance}
\usage{
JC69_distance_allseq(gb, m, adjust_p = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{m}{A matrix of \strong{counts} for bases of the \emph{target} genome to be aligned to bases on the \emph{query} genome.}

\item{adjust_p}{A boolean flag. If \code{TRUE}, the distance is scaled between \code{0} and \code{0.75} to ensure the logarithm stays positive.}
}
\value{
Returns a numeric value representing the evolutionary distance between two genomes. The greater the value, the more genetically different the genomes are.
}
\description{
The Jukes-Cantor 1969 (JC69) distance corrects the p-distance for multiple substitutions,
providing an estimate of evolutionary distance that is proportional to time under the model.
The correction is based on the proportion of nucleotide differences, typically obtained
by counting mismatches in the alignment matrix.
Here, the Jukes-Cantor equation remains unchanged, but the definition of which base pairs
are considered different is modified.
}
\details{
In this function, the fraction of nucleotides that are different incorporates not only the
mismatches from the alignment matrix but also the base pairs that are left unaligned.
The rationale is that these unaligned base pairs likely differ primarily because of point
substitutions and should therefore be treated as mismatches that were not detected by the aligner.
In fact, the aligner can only spot mismatches in regions where the proportion of matches is high enough for alignment.

Notice that gaps are usually not included in the Jukes-Cantor distance, including in this variation,
because they are generally considered to result from indels (insertions and deletions that affect multiple
nucleotides at once), whereas the Jukes-Cantor model is based only on point substitutions.
Therefore, including regions that were likely affected by large evolutionary events, such as gaps caused by indels,
would incorrectly inflate a distance calculated under a model in which only one position is mutated at a time.
}
\examples{

# Only the sequence length is used from the GenomicBreaks object.
gb       <- GRanges(c("Ref:100-35000000:+"))
gb$query <- GRanges(c("Que:1100-35000500:+"))
d <- JC69_distance_allseq(gb, exampleSubstitutionMatrix)

}
\references{
Jukes, T.H. & Cantor, C.R. (1969). "Evolution of protein molecules." In \emph{Mammalian Protein Metabolism} (pp. 21–132). Academic Press.
}
\seealso{
Other Alignment statistics:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GCequilibrium]{GCequilibrium()}},
\code{\link[=GCpressure]{GCpressure()}},
\code{\link[=GCproportion]{GCproportion()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link{exampleSubstitutionMatrix}},
\code{\link[=gapProportion]{gapProportion()}},
\code{\link[=logDet_distance]{logDet_distance()}}

Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Priscila Biller
}
\concept{Alignment statistics}
\concept{Similarity indexes}
