% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepLongestPair.R
\name{keepLongestPair}
\alias{keepLongestPair}
\title{Keep the longest pair of features}
\usage{
keepLongestPair(gb, drop = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object}

\item{drop}{Drop unused sequence levels.}
}
\value{
A \code{GBreaks} object with only one sequence feature in use per genome.
}
\description{
This feature is useful when analysing genomes comprising a main chromosomes
and extra plasmids or smaller chromosomes.  Assuming that the main
chroomosome can be detected only by its length, this function extracts the
alignment between the pair of main chromosomes only.
}
\examples{
exampleTranslocation
exampleTranslocation |> keepLongestPair()

}
\seealso{
Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=chromosome_mapping]{chromosome_mapping()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\author{
Charles Plessy
}
\concept{modifier functions}
