% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{leftInversionGaps}
\alias{leftInversionGaps}
\title{Isolate the left-side gaps in inversions}
\usage{
leftInversionGaps(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges}} object representing the left-side gaps in the
\code{GBreaks} object.
}
\description{
Isolate the left-side gaps in inversions
}
\examples{
leftInversionGaps(flagInversions(exampleInversion))

}
\seealso{
Other Inversion functions:
\code{\link[=filterDoubleInversions]{filterDoubleInversions()}},
\code{\link[=filterInversions]{filterInversions()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=showInversions]{showInversions()}}
}
\concept{Inversion functions}
