% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genomic_breaks.R
\name{load_genomic_breaks}
\alias{load_genomic_breaks}
\title{Load pairwise genome alignments}
\usage{
load_genomic_breaks(
  file,
  target_bsgenome = NULL,
  query_bsgenome = NULL,
  sort = TRUE,
  type = "match_part"
)
}
\arguments{
\item{file}{Path to a file in GFF3 or MAF format.  The file can be compressed
with \emph{gzip}.}

\item{target_bsgenome}{A \code{BSgenome} object or path to a FASTA file representing the \emph{target} genome.}

\item{query_bsgenome}{A \code{BSgenome} object or path to a FASTA file representing the \emph{query} genome.}

\item{sort}{Returns the object sorted, ignoring strand information.}

\item{type}{In GFF3 files, \emph{Sequence Ontology} term representing an alignment
block (default: \code{match_part}).}
}
\value{
Returns a \code{\link{GBreaks}} object where each element represents a pairwise
alignment block.  The \code{granges} part of the object contains the coordinates
on the \emph{target} genome, and the \code{query} metadata column contains the \emph{query}
coordinates in \code{GRanges} format.  The \code{seqinfo} of each \code{BSgenome} object
passed as parameters are copied to the \code{GRanges} objects accordingly.
}
\description{
Loads alignments of a \emph{query} genome to a \emph{target} genome from a text file in
\emph{General Feature Format 3} (GFF3) or \emph{Multiple Alignemnt Format} (MAF).  By
convention, the \emph{target} genome is the one that was indexed by the aligner.
}
\details{
When the input is in GFF3 files, this function expects the pairwise alignment
to be represented in in the following way:
\itemize{
\item Alignments blocks are represented by entries in specific sequence ontology
term in the \emph{type} column.  Other entries will be discarded.  The default
type is \code{match_part}.
\item The coordinate system of the file is the one of the \emph{target} genome.
\item The \code{Target} tag in the \emph{attribute} column contains the coordinates of
the match in the \emph{query} genome.  (Sorry that it is confusing…)
\item Stand information is set so that \emph{query} genome coordinates are always on
the \emph{plus} strand.
}
}
\examples{
load_genomic_breaks(system.file("extdata/contigs.genome.maf.gz", package = "GenomicBreaks"))

\dontrun{
library("BSgenome.Scerevisiae.UCSC.sacCer3")
load_genomic_breaks(
  system.file("extdata/SacCer3__SacPar.gff3.gz", package = "GenomicBreaks"),
  target = Scerevisiae,
  query = NULL)
}

}
\seealso{
The \href{http://www.genome.ucsc.edu/FAQ/FAQformat.html#format5}{MAF format}
documentation on the UCSC genome browser website, and the
\href{https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md}{GFF3 specification}
from the \emph{Sequence Ontology} group.

Other Data loading functions:
\code{\link[=readMAF]{readMAF()}},
\code{\link[=readTrainFile]{readTrainFile()}}
}
\concept{Data loading functions}
