% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_bp_analysis.R
\name{master_bp_analysis}
\alias{master_bp_analysis}
\title{Master Breakpoint Analysis}
\usage{
master_bp_analysis(
  pair_gr,
  co_tol,
  ref_tan,
  ref_tan_tol,
  q_tan,
  q_tan_tol,
  ref_cov,
  ref_cov_tol,
  q_cov,
  q_cov_tol
)
}
\arguments{
\item{pair_gr}{GRanges object containing pairwise alignment}

\item{co_tol}{tolerance over which the coalescing algorithm in \code{coalesce_contigs} should bridge. If left blank, coalescing will not be performed}

\item{ref_tan}{tandem repeat coverage for the reference genome in GRanges format}

\item{ref_tan_tol}{proximity to which a breakpoint should be considered to be "near" a tandem repeat}

\item{q_tan}{tandem repeat coverage for the query genome in GRanges format}

\item{q_tan_tol}{proximity to which a breakpoint should be considered to be "near" a tandem repeat}

\item{ref_cov}{reference genome coverage (per base)}

\item{ref_cov_tol}{range overwhich to give average coverage over a breakpoint (reference genome)}

\item{q_cov}{query genome coverage (per base)}

\item{q_cov_tol}{range overwhich to give average coverage over a breakpoint (query genome)}
}
\value{
a list containing two GRanges objects; one for the reference genome and one for the query genome. Metadata columns contain information per breakpoint, indicated by left (start) breakpoint and right (end) breakpoint
}
\description{
This function brings together the different processes and analyses of the GenomicBreaks package; coalescents, breakpoint coverage and breakpoint proximity to tandem repeats. Input fields can be left blank if the information is not available, and corresponding analyses will not be outputed.
}
