% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oki_cov_pb_data.R
\docType{data}
\name{Oki_cov_pb}
\alias{Oki_cov_pb}
\title{Genome Coverage of Okinawa Genome}
\format{
GRanges object
}
\usage{
Oki_cov_pb
}
\description{
Per-base coverage of the Okinawa Genome (v3). Assembled using BWA.
}
\keyword{datasets}
