% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_distance.R
\name{P_distance}
\alias{P_distance}
\alias{P_error}
\title{Percent difference Distance}
\usage{
P_distance(m, denominator = c("L3", "L1", "L2", "L4"))

P_error(m, denominator = c("L3", "L1", "L2", "L4"))
}
\arguments{
\item{m}{A matrix of counts (\code{P_distance}, \code{P_variance}) or probabilities
(\code{P_distance} only) for bases of the \emph{target} genome to be aligned to bases
on the \emph{query} genome.  As a convenience it can also receive a list produced
by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this matrix.}

\item{denominator}{Denominator according to the nomenclature of May (2004).
Default is \code{L3}.}
}
\value{
A numeric value representing the evolutionary distance between two
genomes.  The larger the value, the more divergent the genomes.
}
\description{
The P-distance is simply the percentage of the aligned base pairs where the
two bases are different.  See Details for a discussion on the many ways to
compute percent identity (and therefore difference.)

The \code{P_error} function reports the standard error given the alignment
length \eqn{n} and percent difference \eqn{p} computed from the matrix:
\eqn{\sqrt{p(1 - p) / n}}.
}
\details{
The P-distance is computed as 1 minus the percent sequence identity
according to the definitions of May (2004), which is the ratio between the
number of matching bases and four possible denominators in use at that time.
\itemize{
\item L1: Length of the shorter sequence.
\item L2: Number of aligned positions, i.e., alignment length (includes gaps, if any).
\item L3: Number of aligned residue pairs, i.e., identities and nonidentities (excludes gaps, if any).
\item L4: Arithmetic mean sequence length.
}
}
\examples{
P_distance(exampleSubstitutionMatrix)

# example code
P_error(exampleSubstitutionMatrix)

}
\references{
May, A. C. W. (2004). Percent Sequence Identity: The Need to Be Explicit.
\emph{Structure}, 12(5), 737–738. \doi{10.1016/j.str.2004.04.001}
}
\seealso{
Other Alignment statistics:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GCequilibrium]{GCequilibrium()}},
\code{\link[=GCpressure]{GCpressure()}},
\code{\link[=GCproportion]{GCproportion()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link{exampleSubstitutionMatrix}},
\code{\link[=gapProportion]{gapProportion()}},
\code{\link[=logDet_distance]{logDet_distance()}}

Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Charles Plessy
}
\concept{Alignment statistics}
\concept{Similarity indexes}
