% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseAlignment.R
\name{pairwiseAlignment}
\alias{pairwiseAlignment}
\alias{pairwiseAlignment,GRanges,GRanges-method}
\alias{pairwiseAlignment,GBreaks,ANY-method}
\title{Pairwise alignment of genomic ranges}
\usage{
pairwiseAlignment(pattern, subject, ...)

\S4method{pairwiseAlignment}{GRanges,GRanges}(pattern, subject, ...)

\S4method{pairwiseAlignment}{GBreaks,ANY}(pattern, subject, ...)
}
\arguments{
\item{pattern}{A \code{GBreaks} or a \code{GRanges} object}

\item{subject}{A \code{GBreaks} or a \code{GRanges} object}

\item{...}{Additional arguments passed to \code{pairwiseAlignment}.}
}
\value{
Returns a \code{\link[pwalign:PairwiseAlignments]{pwalign::PairwiseAlignments}} object.
}
\description{
Retrieves DNA sequence from \code{\link{GBreaks}} or \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges}} objects that are
properly related to a \code{BSgenome} package, and aligns them with the
\code{\link[pwalign:pairwiseAlignment]{pwalign::pairwiseAlignment()}} function.
}
\examples{
Scerevisiae <- BSgenome.Scerevisiae.UCSC.sacCer3::Scerevisiae
# Very arbitrary example
gb <- GBreaks( target = GRanges("chrI: 1-20", seqinfo = seqinfo(Scerevisiae))
             , query  = GRanges("chrI:21-40", seqinfo = seqinfo(Scerevisiae)))
pairwiseAlignment(gb)

}
\seealso{
Other Bioconductor API functions:
\code{\link{GBreaks-class}},
\code{\link[=getSeq]{getSeq()}},
\code{\link[=range_GBreaks]{range_GBreaks()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=subsetByOverlaps_GBreaks]{subsetByOverlaps_GBreaks()}}
}
\author{
Charles Plessy
}
\concept{Bioconductor API functions}
