% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationVector.R
\name{permutationVector}
\alias{permutationVector}
\title{Permutation vector}
\usage{
permutationVector(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
A vector of non-zero integers representing the position of the
genomic content in the \code{GBreaks} object.
}
\description{
Represent \emph{GenomicBreaks} objects as a \emph{permutation vector}.
}
\details{
In a \code{GBreaks} object properly sorted by its \emph{target} ranges, the \emph{query}
ranges …
}
