% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicBreaks.R
\name{range_GBreaks}
\alias{range_GBreaks}
\alias{range,GBreaks-method}
\title{Range method for \code{GBreaks} objects}
\usage{
range_GBreaks(
  x,
  ...,
  with.revmap = FALSE,
  ignore.strand = FALSE,
  na.rm = FALSE
)

\S4method{range}{GBreaks}(x, ..., with.revmap = FALSE, ignore.strand = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{GBreaks} object.}

\item{...}{etc}

\item{with.revmap}{FALSE}

\item{ignore.strand}{FALSE}

\item{na.rm}{FALSE)}
}
\value{
tbd
}
\description{
This is a \code{range} method for \code{\link{GBreaks}} objects, that will run
\code{\link[GenomicRanges:range]{GenomicRanges::range}} on its \emph{target} and \emph{query} ranges and will return
a new \code{GBreaks} object.
}
\note{
\code{range} unconditionally ignores strand in \code{GBreaks} objects.
}
\examples{
range(exampleColinear3)
range(exampleDeletion)
}
\seealso{
Other Bioconductor API functions:
\code{\link{GBreaks-class}},
\code{\link[=getSeq]{getSeq()}},
\code{\link[=pairwiseAlignment]{pairwiseAlignment()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=subsetByOverlaps_GBreaks]{subsetByOverlaps_GBreaks()}}
}
\concept{Bioconductor API functions}
