% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrainFile.R
\name{readTrainFile}
\alias{readTrainFile}
\title{Extract values from \code{last-train} parameter file}
\usage{
readTrainFile(input_file)
}
\arguments{
\item{input_file}{A string of the path to the train file from LAST.}
}
\value{
Returns a \code{list} of parameters from LAST train file. "PercentSimilarity", "PercentSimilarityNogaps", "mean_delete_size", "mean_insert_size", "substitution_percent_identity", "probability",
"matchProb", "delOpenProb", "insOpenProb", "delExtendProb", "insExtendProb", "endProb"
}
\description{
This function is used to read the train file from LAST and extract the final parameters.
}
\examples{
readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))

}
\seealso{
Other Data loading functions:
\code{\link[=load_genomic_breaks]{load_genomic_breaks()}},
\code{\link[=readMAF]{readMAF()}}
}
\author{
Zikun Yang
}
\concept{Data loading functions}
