% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{removeTranslocations}
\alias{removeTranslocations}
\title{Remove translocations}
\usage{
removeTranslocations(gb, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{detect}{Run again \code{flagTranslocations()} if \code{TRUE}, else reuse the \code{tra}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object after removing the ranges that are
translocated and coalescing the resulting collinear ranges.
}
\description{
Remove translocations
}
\examples{
removeTranslocations(exampleTranslocation)

}
\seealso{
Other Translocation functions:
\code{\link[=filterTranslocations]{filterTranslocations()}},
\code{\link[=flagTranslocations]{flagTranslocations()}},
\code{\link[=showTranslocations]{showTranslocations()}}

Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=chromosome_mapping]{chromosome_mapping()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\concept{Translocation functions}
\concept{modifier functions}
