% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Sliding window}
\usage{
slidingWindow(
  gb,
  factor = 100,
  windowSize = NULL,
  stepSize = NULL,
  merged = FALSE,
  cut = TRUE
)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{factor}{Number of windows.}

\item{windowSize}{The size of the window.}

\item{stepSize}{The size of the step.}

\item{merged}{set TRUE to output a \code{GBreaks} object}

\item{cut}{whether cut the alignments out of the window}
}
\value{
Returns a list of \code{GBreaks} objects, one for each window.
}
\description{
This function is used to subset a \code{GBreaks} object into a list of \code{GBreaks}
using a sliding window.
}
\examples{
exdata_Sac <- system.file("extdata/SacCer3__SacPar.gff3.gz", package = "GenomicBreaks")
gb <- load_genomic_breaks(exdata_Sac, BSgenome.Scerevisiae.UCSC.sacCer3::Scerevisiae)
slided_gb <- slidingWindow(gb, windowSize = 1e5, stepSize = 5e4)
slided_gb[[1]]

}
\seealso{
Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=JC69_distance_allseq]{JC69_distance_allseq()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Zhang Kun
}
\concept{Similarity indexes}
