% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSeqLevels.R
\name{splitSeqLevel}
\alias{splitSeqLevel}
\title{Split a seqlevel in two pieces}
\usage{
splitSeqLevel(gb, seq, bp)
}
\arguments{
\item{gb}{A \code{GBreaks} object.}

\item{seq}{A one seqlevel from \code{gr}}

\item{bp}{The breakpoint where to split.}
}
\value{
Returns a modified \code{GRanges} object in which the sequence has been
split  Its \code{\link[Seqinfo:seqinfo]{Seqinfo::seqinfo}} has a new entry for the new levels, and the old
level is not removed.  If no \code{seqlengths} were present in the original
object, they are arbitrarily set as the maximal end value for each \code{seqlevel}.
}
\description{
For scaffolding or plotting purposes, it may be useful to split some
sequences into smaller ones.
}
\note{
This function only splits at breakpoints.
}
\examples{
splitSeqLevel(exampleInsertion, "chrA", 200)

}
\seealso{
Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=swap]{swap()}}

Other scaffolding functions:
\code{\link[=flipStrandNames]{flipStrandNames()}},
\code{\link[=longestMatchesInTarget]{longestMatchesInTarget()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=scaffoldByFlipAndMerge]{scaffoldByFlipAndMerge()}},
\code{\link[=strandNames]{strandNames()}}
}
\author{
Charles Plessy
}
\concept{modifier functions}
\concept{scaffolding functions}
