% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{superhurdles_count}
\alias{superhurdles_count}
\title{Count superhurdles}
\usage{
superhurdles_count(info, g, query_sequence_unsig)
}
\arguments{
\item{info}{Object that stores the hurdles in the breakpoint graph.}

\item{g}{The breakpoint graph.}

\item{query_sequence_unsig}{The extended permutation.}
}
\description{
This function finds superhurdles in the breakpoint graph by checking which hurdles are superhurdles.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
