% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap target and query genomes}
\usage{
swap(gb, sort = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{sort}{Sort the returned object, ignoring strand.}
}
\value{
A new \code{GBreaks} object representing the input's \emph{query} ranges,
with the input's \emph{target} ranges in the \emph{query} slot.  The strand information
is unchanged.
}
\description{
Produce a new object in which the information about the \emph{target} and
\emph{query} genomes have been inverted.  As a large number of \emph{GenomicBreaks}
functions require \code{GBreaks} objects that were sorted with strand information
ignored, a \code{sort} option is provided to do that.
}
\examples{
swap(exampleDeletion)
swap(exampleDeletion, sort = TRUE)

}
\seealso{
Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}}
}
\concept{modifier functions}
