% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntenicMatches.R
\name{syntenicProportions}
\alias{syntenicProportions}
\alias{syntenicMatches}
\title{Sequence levels matched by synteny}
\usage{
syntenicProportions(gb)

syntenicMatches(gb, min = 0.01)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{min}{A proportion threshold between 0 and 1 for a pair of sequence
levels to be considered matching.}
}
\value{
\code{syntenicProportions} returns a list named by the sequence levels
of the \emph{target} genome, indicating what proportion of their aligned sequences
are mapped to a given sequence level of the \emph{query} genome.
\code{syntenicMatches} only retains the names of the \emph{query} sequence levels after
discarding the ones representing a proportion of hits lower than the \code{min}
threshold.
}
\description{
Sequence levels matched by synteny
}
\examples{
syntenicProportions(exampleTranslocation)
syntenicMatches(exampleTranslocation)

}
\seealso{
Other Synteny functions: 
\code{\link{synteny_index}()}
}
\concept{Synteny functions}
