% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny_index.R
\name{synteny_index}
\alias{synteny_index}
\title{Index representing how 'syntenic' two genomes are}
\usage{
synteny_index(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a numeric value between 0 and 1.
}
\description{
A sequence feature on the \emph{target} genome will be considered \sQuote{syntenic}
if most of its alignments map to the same feature of the \emph{query} genome, and
the two genomes are considered syntenic if most of their features are
syntenic.
}
\details{
For a given feature of the \emph{target} genome the number of bases aligned on
each of the features of the \emph{query} genome are computed.  The largest number
is taken.  To make features comparable, this number is then normalised by
total number of aligned bases.  Then, the computed numbers for all the
\emph{target}'s features are averaged with a weighted mean, where the weight is
the total number of aligned bases of that feature.  Thus, a number close to
1 is expected for closely related genomes.
}
\note{
Here, the definition of \sQuote{synteny} is \emph{sequences on the same
DNA strands}, or in technical terms: sequences that are on the same feature,
regardless of strand orientation.
}
\examples{
gb       <- GRanges(c("Ref:100-200:+",   "Ref:400-500:+",    "Ref:600-700:+"))
gb$query <- GRanges(c("Que:1100-1200:+", "Que2:1700-1800:+", "Que:1500-1600:+"))
synteny_index(gb)

}
\seealso{
Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\concept{Similarity indexes}
