% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tan_bp.R
\name{tan_bp}
\alias{tan_bp}
\title{Breakpoint proximity to Tandem Repeats}
\usage{
tan_bp(gr_ob, tan, tol, query_tf)
}
\arguments{
\item{gr_ob}{GRanges object containing pairwise alignment}

\item{tan}{GRanges object containing tandem repeat (or characterisitc of interest) coverage}

\item{tol}{range in which breakpoints will be classified as "near" to a tandem repeat}

\item{query_tf}{logical value that, if set to TRUE, will classify bps in query coordinates as opposed to reference (default)}
}
\value{
A list containing 3 GRanges objects of breakpoints far, near and within tandem repeats.
}
\description{
This function classifies breakpoints in terms of its proximity to tandem repeats (but can be generalised to any property whos coverage is binary).
}
