% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tan_O_data.R
\docType{data}
\name{tan_O}
\alias{tan_O}
\title{Tandem Repeat coverage of the Osaka2016 Genome}
\format{GRanges object}
\usage{
tan_O
}
\description{
Per-base tandem repeat coverage of the Osaka2016 genome. Tantan was used to determine the coverage.
}
\keyword{datasets}
