% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tan_Oki_data.R
\docType{data}
\name{tan_Oki}
\alias{tan_Oki}
\title{Tandem Repeat coverage of the Okinawa (v3) Genome}
\format{GRanges object}
\usage{
tan_Oki
}
\description{
Per-base tandem repeat coverage of the Okinawa (v3) genome. Tantan was used to determine the coverage.
}
\keyword{datasets}
