% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem_coverage.R
\name{tandem_coverage}
\alias{tandem_coverage}
\title{Tandem Repeat Coverage Heatmap}
\usage{
tandem_coverage(gr_ob, tan_ref, win, lab)
}
\arguments{
\item{gr_ob}{GRanges object containing pairwise alignment}

\item{tan_ref}{GRanges object containing coverage of tandem repeat in reference genome. To aquire the analysis for the query genome, reformat the GRanges object to have the query as the main subject, and replace \code{tan_ref} with the query genome tandem coverage}

\item{win}{window over which to observe tandem coverage over each breakpoint}

\item{lab}{label for the plot}
}
\value{
The output is a coverage heatmap that can be plotted with smoothHeatmap(output), then pipe into plotHeatmapLis
}
\description{
A function that produces a heatmap of the tandem coverage around breakpoints. This can be generalised to any characteristic which displays binary coverage.
}
