% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{bp_count}
\alias{bp_count}
\title{Number of Breakpoints}
\usage{
bp_count(p_extended)
}
\arguments{
\item{p_extended}{The extended permutation.}
}
\value{
The number of breakpoints between two genomes.
}
\description{
This function computes the number of breakpoints given an extended permutation.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=cycle_nontrivial_count]{cycle_nontrivial_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
