% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_heatmap.R
\name{bp_heatmap}
\alias{bp_heatmap}
\title{Breakpoint Associated Heatmaps for Pairwise Aligned Genomes}
\usage{
bp_heatmap(gr_ob, gen_seq, basep_range, pat)
}
\arguments{
\item{gr_ob}{GRanges object containing pairwise alignment}

\item{gen_seq}{This should be either a BSgenome object or a DNAStringSet object, such that the genome sequnce is contained in this input}

\item{basep_range}{range overwhich heatmap characteristic is plotted. Breakpoints will be aligned at the center of this.}

\item{pat}{character string of desired pattern/characteristic to be plotted on heatmap}
}
\value{
Heatmap of pattern around centred breakpoints
}
\description{
This function generates a Heatmap of the specificed pattern, over breakpoints aligned at the centre of the plot. The result can be piped into \code{smoothHeatmap} and then \code{plotHeatmapList} or \code{plotHeatmapMeta}
}
