% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_pair_analysis.R
\name{bp_pair_analysis}
\alias{bp_pair_analysis}
\title{Breakpoint Pairwise Comparison using Reference Genome as Coordinate System}
\usage{
bp_pair_analysis(gr_ref_q1, gr_ref_q2, win, lab)
}
\arguments{
\item{gr_ref_q1}{pairwise alignment between the reference genome and query genome 1}

\item{gr_ref_q2}{is the pw alignment between the reference genome and query genome 2}

\item{win}{range around query1 breakpoints of which query2 breakpoints are projected on to. Should be even number}

\item{lab}{string which will be the label on the graph}
}
\value{
The output is a coverage heatmap that can be plotted with smoothHeatmap(output), then pipe into plotHeatmapList
}
\description{
This function takes in two pairwise alignments, with the same reference genome. A heatmap is produced, where the breakpoints of query1 are the centre point reference, and the breakpoints of query2 that fall within the window are projected onto it. Format of the pairwise alignments are expected to be GRanges objects, with information about the query genome in the metadata column.
}
