% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_pair_analysis.R
\name{bp_pair_analysis}
\alias{bp_pair_analysis}
\title{Breakpoint Pairwise Comparison using Reference Genome as Coordinate System}
\usage{
bp_pair_analysis(gb1, gb2, window, label)
}
\arguments{
\item{gb1}{\code{GBreaks} object of the alignment between the \emph{target} genome and \emph{query} genome 1.}

\item{gb2}{\code{GBreaks} object of the alignment between the \emph{target} genome and \emph{query} genome 2.}

\item{window}{range around query1 breakpoints of which query2 breakpoints are projected on to. Should be even number}

\item{label}{string which will be the label on the graph}
}
\value{
Returns a \code{\link{Heatmap}} object of \code{pattern} that can be piped into
\code{\link[heatmaps:smoothHeatmap]{heatmaps::smoothHeatmap}} and then \code{\link[heatmaps:plotHeatmapList]{heatmaps::plotHeatmapList}} or
\code{\link[heatmaps:plotHeatmapMeta]{heatmaps::plotHeatmapMeta}}.
}
\description{
This function takes two \code{\link{GBreaks}} objects, with the same \emph{target} genome.
A heatmap is produced, where the breakpoints of \emph{query} 1 are the centre
point reference, and the breakpoints of \emph{query} 2 that fall within the window
are projected onto it.
}
\seealso{
Other plot functions: 
\code{\link{bp_heatmap}()},
\code{\link{feature_coverage}()},
\code{\link{makeOxfordPlots}()},
\code{\link{plotApairOfChrs}()}

Other heatmap functions: 
\code{\link{bp_heatmap}()},
\code{\link{feature_coverage}()}
}
\author{
Charlotte West
}
\concept{heatmap functions}
\concept{plot functions}
