% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{breakpoint_graph}
\alias{breakpoint_graph}
\title{Breakpoint Graph}
\usage{
breakpoint_graph(p_extended)
}
\arguments{
\item{p_extended}{The extended permutation.}
}
\value{
The breakpoint graph.
}
\description{
This function creates a breakpoint graph given an extended permutation.
}
\references{
Hannenhalli, Sridhar, and Pavel A. Pevzner. "Transforming cabbage into turnip: polynomial algorithm for sorting signed permutations by reversals." Journal of the ACM (JACM) 46.1 (1999): 1-27.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=cycle_nontrivial_count]{cycle_nontrivial_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
