% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{breakpointGraphProperties}
\alias{breakpointGraphProperties}
\title{Breakpoint graph properties}
\usage{
breakpointGraphProperties(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
A list containing the following properties of the breakpoint graph:
1. \code{N} : the total number of places where a breakpoint could occur (which is the same as the \code{number_aligned_blocks + 1});
2. \code{nbBreakpoints} : the total number of breakpoints in the extended permutation;
3. \code{nbCycles} : the total number of cycles in the breakpoint graph.
}
\description{
Computes the breakpoint graph of a pair of chromosomes
and returns key properties of the graph, which are used
in various rearrangement estimates.
}
\references{
Hannenhalli, Sridhar, and Pavel A. Pevzner. "Transforming cabbage into turnip: polynomial algorithm for sorting signed permutations by reversals." Journal of the ACM (JACM) 46.1 (1999): 1-27.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}

Other Rearrangement distances:
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}}

Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Priscila Biller
}
\concept{Breakpoint graph functions}
\concept{Rearrangement distances}
\concept{Similarity indexes}
