% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_contigs.R
\name{chain_contigs}
\alias{chain_contigs}
\title{Chain contigs in pairwise alignments}
\usage{
chain_contigs(gb, tol = Inf, drop.within = FALSE)
}
\arguments{
\item{gb}{\code{\link{GBreaks}} object of the pairwise alignment.}

\item{tol}{Unaligned region of width lesser than or equal to \code{tol} in both
the reference and query case will be bridged in coalescing.}

\item{drop.within}{Discard the chains that are included in other chains.}
}
\value{
TBW
}
\description{
Description TBW
}
\examples{

gb4       <- GenomicRanges::GRanges(c("Ref:100-200:+", "Ref:400-500:+", "Ref:600-700:+"))
gb4$query <- GenomicRanges::GRanges(c("Que:1100-1200:+", "Que:1700-1800:+", "Que:1500-1600:+"))
chain_contigs(gb4)

}
\seealso{
Other Colinearity functions: 
\code{\link{coalesce_contigs}()},
\code{\link{flagColinearAlignments}()}
}
\author{
Charles Plessy
}
\concept{Colinearity functions}
