% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_contigs.R
\name{chain_contigs}
\alias{chain_contigs}
\title{Chain contigs in pairwise alignments}
\usage{
chain_contigs(gb, tol = Inf, drop.within = FALSE)
}
\arguments{
\item{gb}{\code{\link{GBreaks}} object of the pairwise alignment.}

\item{tol}{Unaligned region of width lesser than or equal to \code{tol} in both
the reference and query case will be bridged in coalescing.}

\item{drop.within}{Discard the chains that are included in other chains.}
}
\value{
TBW
}
\description{
Description TBW
}
\examples{
chain_contigs(exampleInversion)

}
\seealso{
Other Colinearity functions: 
\code{\link{GOC}()},
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{dist2next}()},
\code{\link{filterColinearRegions}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagPairs}()}
}
\author{
Charles Plessy
}
\concept{Colinearity functions}
