% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGaps.R
\name{cleanGaps}
\alias{cleanGaps}
\title{Gaps between ranges}
\usage{
cleanGaps(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} or a \code{\link{GRanges}} object.}
}
\value{
Returns a strandless \code{\link{GRanges}} object representing the gaps
between the ranges of the input.
}
\description{
Utility function that runs \code{GenomicRanges::gaps()} and then cleans its output
by removing strandless ranges as well as the last range that represents the
sequence between the end of the input object and the end of its sequence
levels.
}
\note{
If you find replacement provided by a package that we already import,
please let me know in a GitHub issue or pull request.
}
\examples{
cleanGaps(exampleColinear)
GenomicRanges::gaps(exampleColinear)
cleanGaps(exampleInversion)

}
\author{
Charles Plessy
}
