% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_contigs.R
\name{coalesce_contigs}
\alias{coalesce_contigs}
\title{Algorithm for Coalescing Pairwise Alignments}
\usage{
coalesce_contigs(gr_ob, tol)
}
\arguments{
\item{gr_ob}{GRanges object of the pairwise alignment, with reference genome as the subject of the GRanges object, and query genome alignment in the metadata column "name".}

\item{tol}{width of gap that will be bridged in coalescing. The gap must be less than or equal to \code{"tol"} in both the reference and query case.}
}
\value{
new GRanges object of similar structure (GRanges main object is reference genome and metadata is query genome) but with a reduced number of alignment fragments due to coalescion.
}
\description{
This algorithm take in pairwise alignment, and reduces the number of alignments by coalescing fragments that are within close proximity (user determined). Fragmented alignments can cause artificial breakpoints arising from incorrect basecalls, misassembly and misalignments.
}
