% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_contigs.R
\name{coalesce_contigs}
\alias{coalesce_contigs}
\title{Coalesce Pairwise Alignments}
\usage{
coalesce_contigs(gb, tol = Inf, minwidth = 0)
}
\arguments{
\item{gb}{\code{\link{GBreaks}} object of the pairwise alignment.}

\item{tol}{Unaligned region of width lesser than or equal to \code{tol} in both
the reference and query case will be bridged in coalescing.}

\item{minwidth}{Remove the intervals whose width smaller than this value.}
}
\value{
Returns a new \code{GBreaks} object with a reduced number of alignments
fragments due to coalescion.  The returned object is sorted ignoring strand.
For convenience during analysis sessions, its \emph{score} is set to the width of
the ranges on the \emph{target} genome.
}
\description{
Coalesce Pairwise Alignments
}
\note{
Fragmented alignments arising from incorrect basecalls, misassembly or
misalignments can cause us to infer artificial breakpoints

Internally, \code{coalesce_contigs} uses \code{\link[=flagColinearAlignments]{flagColinearAlignments()}}  See the
examples for details.
}
\examples{
flagColinearAlignments(exampleColinear3, details = TRUE)
coalesce_contigs(exampleColinear3)

# Target range [1] precedes target range [2]
precede(exampleColinear3)
# Query range [1] precedes query range [2]
precede(exampleColinear3$query)

# Ranges on the minus strand
gb2 <- exampleColinear3 |> reverse() |> sort(ignore.strand = TRUE)
flagColinearAlignments(gb2, details = TRUE)
coalesce_contigs(gb2)

# Target range [1] follows target range [2]
follow(gb2)
# Or, ignoring strand, target range [1] precedes target range [2]
precede(gb2, ignore.strand = TRUE)
# Query range [1] follows query range [2]
follow(gb2$query)

# Coalescing strandless objects
gb3 <- exampleColinear3
gb4 <- gb2
strand(gb4) <- strand(gb3) <- "*"
coalesce_contigs(gb3)
coalesce_contigs(gb4)

}
\seealso{
Other Colinearity functions: 
\code{\link{GOC}()},
\code{\link{bridgeRegions}()},
\code{\link{chain_contigs}()},
\code{\link{dist2next}()},
\code{\link{filterColinearRegions}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagPairs}()}

Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{forceSeqLengths}()},
\code{\link{guessSeqLengths}()},
\code{\link{mergeSeqLevels}()},
\code{\link{reverse}()},
\code{\link{swap}()}
}
\author{
Charlotte West

Charles Plessy
}
\concept{Colinearity functions}
\concept{modifier functions}
