% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{components_graph}
\alias{components_graph}
\title{Connected components}
\usage{
components_graph(g, p_extended)
}
\arguments{
\item{g}{The breakpoint graph.}

\item{p_extended}{The extended permutation.}
}
\value{
Connected components of the breakpoint graph.
}
\description{
Connected components of the breakpoint graph.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=cycle_nontrivial_count]{cycle_nontrivial_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
