% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_index.R
\name{correlation_index}
\alias{correlation_index}
\title{Index representing how correlated syntenic regions are}
\usage{
correlation_index(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a numeric value between 0 and 1.
}
\description{
A sequence feature on the \emph{target} genome is considered \sQuote{syntenic}
to the feature of the \emph{query} genome matched by the most of its alignments.
The correlation coefficient of the center of syntenic alignments is then
calculated.  Then, the computed numbers for all the \emph{target}'s features are
averaged as a mean weighted by is the total number of aligned bases of each
feature.  Thus, a number close to 1 is expected for closely related genomes.
}
\note{
Here, the definition of \sQuote{synteny} is \emph{sequences on the same
DNA strands}, or in technical terms: sequences that are on the same feature,
regardless of strand orientation.
}
\examples{
gb       <- GRanges(c("Ref:100-200:+",   "Ref:400-500:+",    "Ref:600-700:+"))
gb$query <- GRanges(c("Que:1100-1200:+", "Que2:1700-1800:+", "Que:1500-1600:+"))
correlation_index(gb)

}
\seealso{
Other Similarity indexes: 
\code{\link{F81_distance}()},
\code{\link{GOC}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{P_distance}()},
\code{\link{T92_distance}()},
\code{\link{karyotype_index}()},
\code{\link{slidingWindow}()},
\code{\link{strand_randomisation_index}()},
\code{\link{synteny_index}()},
\code{\link{tau_index}()}
}
\concept{Similarity indexes}
