% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{cycle_count}
\alias{cycle_count}
\title{Number of Cycles}
\usage{
cycle_count(g)
}
\arguments{
\item{g}{The breakpoint graph.}
}
\value{
The number of cycles in a breakpoint graph.
}
\description{
This function computes the number of cycles (trivial and non-trivial) in a breakpoint graph.
Non-trivial cycles are cycles with more than 2 vertices, whereas trivial cycles are cycles with 2 vertices that represent common adjacencies.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_nontrivial_count]{cycle_nontrivial_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Priscila Biller
}
\concept{Breakpoint graph functions}
