% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2next.R
\name{dist2next}
\alias{dist2next}
\title{Distances to next ranges}
\usage{
dist2next(gr_ob, step = 1, ignore.strand = FALSE)
}
\arguments{
\item{gr_ob}{A GenomicBreaks object.}

\item{step}{Distance to the n^th^ block (default: first).}

\item{ignore.strand}{Calculate distance for ranges on different starands.}
}
\value{
Returns the object with two extra metadata colums, \code{tdist} and
\code{qdist}, containing the distance to the next range in the reference and query
genomes respectively.
}
\description{
Calculates the distance to the next range in the reference and query genome.
}
\details{
Note that because the last range has not next neighbor, the last value is set
to Inf arbitrarily.
}
\examples{
gb1       <- GenomicRanges::GRanges(c("chr1:100-200:+", "chr1:201-300:+", "chr1:400-500:+", "chr1:700-800:-"))
gb1$query <- GenomicRanges::GRanges(c("chr1:100-200:+", "chr1:201-300:+", "chr1:400-500:+", "chr1:700-800:-"))
dist2next(gb1)
dist2next(gb1, ignore.strand = TRUE)
dist2next(gb1, 2)

}
