% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSubstitutionMatrix.R
\docType{data}
\name{exampleSubstitutionMatrix}
\alias{exampleSubstitutionMatrix}
\title{Example Substitution Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
exampleSubstitutionMatrix
}
\description{
Loaded from the file \code{extdata/example.train} distributed in this package.
}
\keyword{datasets}
