% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{extendedPermutation}
\alias{extendedPermutation}
\title{Extended Permutation}
\usage{
extendedPermutation(p)
}
\arguments{
\item{p}{A signed permutation vector.}
}
\value{
The extended permutation.
}
\description{
This function transforms a permutation with \code{N} elements into a permutation with \verb{2N} elements.
The idea is that each element in the input permutation corresponds to an aligned block (or gene, conserved region, etc.).
This function creates two elements for each original element: one representing the start of the block and the other representing the end.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
