% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F81_distance.R
\name{F81_distance}
\alias{F81_distance}
\title{Felsenstein's 1981 Distance}
\usage{
F81_distance(train_parameters)
}
\arguments{
\item{train_parameters}{A list containing the probabilities of the alignment,
produced by the \code{\link[=readTrainFile]{readTrainFile()}} function.}
}
\value{
A numeric value representing the evolutionary distance between two
genomes.  The larger the value, the more divergent the genomes.
}
\description{
The Felsenstein 1981 (F81) distance corrects for multiple substitutions like
the \code{\link[=JC69_distance]{JC69_distance()}} function, but also accounts for GC content. It is
calculated as \eqn{-E \ln\left(\frac{E - p}{E}\right)}, where \eqn{E} equals
\eqn{1} minus the sum of the squares of the average nucleotide frequencies
(McGuire and coll., 1999).
}
\examples{
parameters <- readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))
F81_distance(parameters)

}
\references{
Felsenstein, J. (1981). Evolutionary trees from DNA sequences: a maximumlikelihood approach.
\emph{Journal of Molecular Evolution}, 17, 368–376. \doi{10.1007/BF01734359}

McGuire, G., Prentice, M. J., & Wright, F. (1999). Improved error bounds for genetic distances from DNA sequences.
\emph{Biometrics}, 55(4), 1064–1070. \doi{10.1111/j.0006-341x.1999.01064.x}
}
\seealso{
Other Similarity indexes: 
\code{\link{GOC}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{T92_distance}()},
\code{\link{correlation_index}()},
\code{\link{karyotype_index}()},
\code{\link{slidingWindow}()},
\code{\link{strand_randomisation_index}()},
\code{\link{synteny_index}()}
}
\author{
Charles Plessy
}
\concept{Similarity indexes}
