% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_coverage.R
\name{feature_coverage}
\alias{feature_coverage}
\title{Feature coverage heatmap}
\usage{
feature_coverage(gr_ob, feat, win, lab, ...)
}
\arguments{
\item{gr_ob}{GenomicBreaks object containing pairwise alignment}

\item{feat}{GRanges object containing the feature in target genome.}

\item{win}{window over which to observe tandem coverage over each breakpoint}

\item{lab}{label for the plot}

\item{...}{Other arguments passed to \code{get_bps}.}
}
\value{
The output is \code{Heatmap} object representing a coverage heatmap that
can be postprocessed with \code{smoothHeatmap(output)}, then piped into
\code{plotHeatmapList}.
}
\description{
Produces a heatmap of a genomic feature around breakpoints. The feature has
binary coverage represented by a genomic range.
}
