% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_coverage.R
\name{feature_coverage}
\alias{feature_coverage}
\title{Feature coverage heatmap}
\usage{
feature_coverage(gr, feat, window, label, ...)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{feat}{A \code{GRanges} object representing the feature of interest.}

\item{window}{Window over which to observe feature coverage.}

\item{label}{Label for the plot.}

\item{...}{Other arguments passed to \code{\link{get_bps}} in order to select the
boundaries, their order and their orientation.}
}
\value{
Returns a \code{\link{Heatmap}} object that can be piped into
\code{\link[heatmaps:smoothHeatmap]{heatmaps::smoothHeatmap}} and then \code{\link[heatmaps:plotHeatmapList]{heatmaps::plotHeatmapList}} or
\code{\link[heatmaps:plotHeatmapMeta]{heatmaps::plotHeatmapMeta}}.
}
\description{
Wraps the \code{\link[heatmaps:CoverageHeatmap]{heatmaps::CoverageHeatmap}} function to produce a heatmap centred
on the boundaries of \emph{genomic ranges}.  Assuming that these ranges are a
\code{\link{GBreaks}} object, then the boundaries approximate breakpoints.
}
\seealso{
Other plot functions: 
\code{\link{bp_heatmap}()},
\code{\link{bp_pair_analysis}()},
\code{\link{makeOxfordPlots}()},
\code{\link{plotApairOfChrs}()}

Other heatmap functions: 
\code{\link{bp_heatmap}()},
\code{\link{bp_pair_analysis}()}
}
\concept{heatmap functions}
\concept{plot functions}
