% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{filterDoubleInversions}
\alias{filterDoubleInversions}
\title{Extract central blocks in double inversions}
\usage{
filterDoubleInversions(gb, rename = TRUE, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{rename}{Replace range names by their numeric order before subsetting.}

\item{detect}{Run again \code{flagDoubleInversions()} if \code{TRUE}, else reuse the \code{Dbl}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
the central part of a double inversion quintuplet have been discarded.  If
the object was missing the \code{Dbl} metadata column, return the object after
discarding all of its ranges.
}
\description{
Extract central blocks in double inversions
}
\examples{
filterDoubleInversions(exampleDoubleInversion1)

}
\seealso{
Other Inversion functions: 
\code{\link{filterInversions}()},
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flagTwinInversions}()},
\code{\link{flipInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{removeDoubleInversions}()},
\code{\link{removeInversions}()},
\code{\link{showInversions}()}
}
\concept{Inversion functions}
