% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{filterDoubleInversions}
\alias{filterDoubleInversions}
\title{Extract central blocks in double inversions}
\usage{
filterDoubleInversions(gb, rename = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object processed with \code{\link{flagDoubleInversions}}.}

\item{rename}{Replace range names by their numeric order before subsetting.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
the central part of a double inversion quintuplet have been discarded.  If
the object was missing the \code{Dbl} metadata column, return the object after
discarding all of its ranges.
}
\description{
Extract central blocks in double inversions
}
\examples{
filterDoubleInversions(flagDoubleInversions(exampleDoubleInversion1))

}
\seealso{
Other Inversion functions: 
\code{\link{filterInversions}()},
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flipInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{showInversions}()}
}
\concept{Inversion functions}
