% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{filterInversions}
\alias{filterInversions}
\title{Extract central blocks in inversions}
\usage{
filterInversions(gb, rename = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object processed with \code{\link{flagInversions}}.}

\item{rename}{Replace range names by their numeric order before subsetting.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
the central part of an inversion triplet have been discarded.  If the object
was missing the \code{inv} metadata column, return the object after discarding all
of its ranges.
}
\description{
Extract central blocks in inversions
}
\examples{
filterInversions(flagInversions(exampleInversion))

}
\seealso{
Other Inversion functions: 
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{showInversions}()}
}
\concept{Inversion functions}
