% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{filterTranslocations}
\alias{filterTranslocations}
\title{Extract central blocks in translocations}
\usage{
filterTranslocations(gb, rename = TRUE, remove = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object processed with \code{\link{flagTranslocations}}.}

\item{rename}{Replace range names by their numeric order before subsetting.}

\item{remove}{Filter out instead of filtering in.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
the central part of an inversion triplet have been discarded.  If the object
was missing the \code{tra} metadata column, return the object after discarding all
of its ranges.
}
\description{
Extract central blocks in translocations
}
\examples{
filterTranslocations(flagTranslocations(exampleTranslocation))
filterTranslocations(remove = TRUE, flagTranslocations(exampleTranslocation))

}
\seealso{
Other Translocation functions: 
\code{\link{flagTranslocations}()},
\code{\link{showTranslocations}()}
}
\concept{Translocation functions}
