% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{filterTranslocations}
\alias{filterTranslocations}
\title{Extract central blocks in translocations}
\usage{
filterTranslocations(gb, rename = TRUE, remove = FALSE, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{rename}{Replace range names by their numeric order before subsetting.}

\item{remove}{Filter out instead of filtering in.}

\item{detect}{Run again \code{flagTranslocations()} if \code{TRUE}, else reuse the \code{tra}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
the central part of an inversion triplet have been discarded.  If the object
was missing the \code{tra} metadata column, return the object after discarding all
of its ranges.
}
\description{
Extract central blocks in translocations
}
\examples{
filterTranslocations(exampleTranslocation)
filterTranslocations(remove = TRUE, exampleTranslocation)

}
\seealso{
Other Translocation functions: 
\code{\link{flagTranslocations}()},
\code{\link{removeTranslocations}()},
\code{\link{showTranslocations}()}
}
\concept{Translocation functions}
