% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagAll.R
\name{flagAll}
\alias{flagAll}
\title{Compute all flags}
\usage{
flagAll(gb)
}
\arguments{
\item{gb}{A \code{GenomicBreaks} object.}
}
\value{
Returns the \code{GenomicBreaks} object with an extra \code{flag} metadata
column.
}
\description{
Flag ranges that are at the beginning of a colinear duplet, or an inversion,
insertion or interruption triplet.
}
\examples{
inv       <- GRanges(c("XSR:101-200:+", "XSR:201-300:-",  "XSR:301-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
flagAll(inv)

}
\seealso{
Other Flagging functions: 
\code{\link{flagColinearAlignments}()},
\code{\link{flagInsersions}()},
\code{\link{flagInterruptions}()},
\code{\link{flagInversions}()},
\code{\link{flagLongShort}()}
}
\concept{Flagging functions}
