% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagColinearAlignments.R
\name{flagColinearAlignments}
\alias{flagColinearAlignments}
\title{Flag colinear alignments}
\usage{
flagColinearAlignments(gb, tol = Inf, details = FALSE)
}
\arguments{
\item{gb}{\code{\link{GBreaks}} object of the pairwise alignment.}

\item{tol}{Unaligned regions larger than this \emph{tolerance} threshold will
interrupt colinearity.}

\item{details}{Returns more metadata columns if \code{TRUE}}
}
\value{
Returns a modified \code{GBreaks} object with a new \code{colinear} metadata
column indicating if an alignment is colinear with the next one.  If the
\code{details} option is set to \code{TRUE}, it will also output the relative position
of the following and previous alignments on the \emph{target} and \emph{query} genomes
(\code{tfoll}, \code{tprev}, \code{qfoll}, \code{qprev}), a partial colinearity flag for each
genome (\code{t_col} and \code{q_col}), and the distance to the next alignment on
each genome (\code{tdist} and \code{qdist}).
}
\description{
Flags alignments that are colinear with the next one in sequence order.
The flag is added to the first alignment.  Colinearity is defined by the fact
that the next alignment on the \emph{target} genome is the same as the next
alignment on the \emph{query} genome, with strand information being taken into
account.
}
\details{
Internally, \code{flagColinearAlignments()} uses the \code{\link[GenomicRanges:nearest-methods]{GenomicRanges::precede}}
and \code{\link[GenomicRanges:nearest-methods]{GenomicRanges::follow}} functions functions to determine what is the
\emph{next}.  For a given range, these functions return the index position of the
range it precedes or follows, or \code{NA} as the first range follows nothing and
the last range precedes nothing.  See the examples for details.
}
\note{
The flags are only valid as long as the Genomic Breaks object is not
modified by removing alignments or sorting them in a different order.
}
\examples{
# Ranges on the plus strand that should coalesce
gb1       <- GenomicRanges::GRanges(c(A="Ref:100-200:+", B="Ref:400-500:+"))
gb1$query <- GenomicRanges::GRanges(c(A="Que:100-200",   B="Que:400-500"))
gb1
flagColinearAlignments(gb1)

# Reference range [1] precedes reference range [2]
GenomicRanges::precede(gb1)
# Query range [1] precedes query range [2]
GenomicRanges::precede(gb1$query)

# Ranges on the minus strand that should coalesce
gb2       <- GenomicRanges::GRanges(c(B="Ref:100-200:-", A="Ref:400-500:-"))
gb2$query <- GenomicRanges::GRanges(c(B="Que:400-500",   A="Que:100-200"))
gb2$qname <- names(gb2$query)
gb2
# Reference range [1] follows reference range [2]
GenomicRanges::follow(gb2)
# Or, ignoring strand, reference range [1] precedes reference range [2]
GenomicRanges::precede(gb2, ignore.strand = TRUE)
# Query range [1] follows query range [2]
GenomicRanges::follow(gb2$query)
flagColinearAlignments(gb2)

# Ranges on the minus strand that should not coalesce because they are not
# ordered properly
gb3       <- GenomicRanges::GRanges(c("Ref:100-200:-", "Ref:400-500:-"))
gb3$query <- GenomicRanges::GRanges(c("Que:100-200",   "Que:400-500"))
# Reference range [1] follows reference range [2]
GenomicRanges::follow(gb3)
# Query range [1] follows query range [2]
GenomicRanges::precede(gb3$query)
flagColinearAlignments(gb3)

Ranges on the plus strand that should not coalesce because they are not
ordered properly
gb4       <- GenomicRanges::GRanges(c("Ref:100-200:+", "Ref:400-500:+", "Ref:600-700:+"))
gb4$query <- GenomicRanges::GRanges(c("Que:1100-1200:+", "Que:1700-1800:+", "Que:1500-1600:+"))
flagColinearAlignments(gb4)

}
\seealso{
Other Flagging functions: 
\code{\link{flagInsersions}()},
\code{\link{flagInversions}()}

Other Colinearity functions: 
\code{\link{coalesce_contigs}()}
}
\author{
Charlotte West

Charles Plessy
}
\concept{Colinearity functions}
\concept{Flagging functions}
