% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flagDoubleInversions}
\alias{flagDoubleInversions}
\title{Flag double inversions}
\usage{
flagDoubleInversions(gb, details = FALSE)
}
\arguments{
\item{gb}{A \code{\link[=GBreaks]{GBreaks()}} object.}

\item{details}{Returns more metadata columns if \code{TRUE}.}
}
\description{
Two consecutive and overlapping inversions will generate patterns that can
this function aims to detect.
}
\examples{
# Start colinear.  Lower case meands minus strand
z0 <- GBreaks(target = GRanges(c(A="T:10-15:+", B="T:20-25:+", C="T:30-35:+")),
              query  = GRanges(c(A="Q:10-15",   B="Q:20-25",   C="Q:30-35")))

# Swap coordinates of A and B on the query and flip strands  ABC -> baC
z1 <- GBreaks(target = GRanges(c(A="T:10-15:-", B="T:20-25:-", C="T:30-35:+")),
              query  = GRanges(c(a="Q:20-25",   b="Q:10-15",   C="Q:30-35")))

# Now query order is b - a - C.  Swap a and C and flip strands baC -> bcA
z2 <- GBreaks(target = GRanges(c(A="T:10-15:+", B="T:20-25:-", C="T:30-35:-")),
              query  = GRanges(c(A="Q:30-35",   b="Q:10-15",   c="Q:20-25")))

# Altogether, there are:
# ABC -> baC -> bcA
# ABC -> Acb -> Cab
# cba -> BCa -> BAc
# cba -> cAB -> aCB


# z2 has same topoloty as package example
plotApairOfChrs(z2)
plotApairOfChrs(exampleDoubleInversion1)

flagDoubleInversions(exampleDoubleInversion1 )

exampleDoubleInversion1Rev <- reverse(exampleDoubleInversion1) |> sort(ignore.strand = TRUE)
exampleDoubleInversion1Rev[2:4] |> plotApairOfChrs()

}
\seealso{
Other Flagging functions: 
\code{\link{flagAll}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagInversions}()},
\code{\link{flagLongShort}()},
\code{\link{flagPairs}()},
\code{\link{flagTranslocations}()}

Other Inversion functions: 
\code{\link{filterDoubleInversions}()},
\code{\link{filterInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flipInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{removeInversions}()},
\code{\link{showInversions}()}

Other Structural variants: 
\code{\link{StructuralVariants}},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flagTranslocations}()},
\code{\link{plotApairOfChrs}()}
}
\author{
Charles Plessy
}
\concept{Flagging functions}
\concept{Inversion functions}
\concept{Structural variants}
