% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flagInversions}
\alias{flagInversions}
\title{Flag inversions}
\usage{
flagInversions(gb, tol = Inf)
}
\arguments{
\item{gb}{A \code{GBreaks()} object.}

\item{tol}{Tolerance window for the distance between two ranges.}
}
\value{
Returns the \code{GBreaks} object with an extra \code{inv} metadata column.
}
\description{
Flag ranges that start a triplet that would be colinear if the central pair
were aligned to the opposite strand.
}
\details{
Here is a trivial example of an inversion.\preformatted{┌──────────────┬──────────────┬──────────────┐
│ chrA:101-200 │ chrA:201-300 │ chrA:301-400 │ (Target genome)
└──────────────┴──────────────┴──────────────┘
       +               -             +         (Alignment direction)
┌──────────────┬──────────────┬──────────────┐
│ chrB:101-200 │ chrB:201-300 │ chrB:301-400 │ (Query genome)
└──────────────┴──────────────┴──────────────┘
}
}
\examples{
flagInversions(exampleInversion)
plotApairOfChrs(exampleInversion)

# Not an inversion
flagInversions(exampleTranslocation)
inv3       <- GRanges(c("XSR:101-180:-", "chr1:201-300:+",  "XSR:320-400:-"))
inv3$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
any(flagInversions(inv3)$inv) == FALSE

# Also not an inversion
inv4       <- GRanges(c("XSR:101-180:-", "XSR:1201-1300:+", "XSR:320-400:-"))
inv4$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
flagInversions(inv4)

any(flagInversions(inv4)$inv) == FALSE


}
\seealso{
Other Flagging functions: 
\code{\link{flagAll}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagInsersions}()},
\code{\link{flagLongShort}()}

Other Inversion functions: 
\code{\link{filterInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{showInversions}()}

Other Structural variants: 
\code{\link{StructuralVariants}},
\code{\link{plotApairOfChrs}()}
}
\concept{Flagging functions}
\concept{Inversion functions}
\concept{Structural variants}
