% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flagInversions}
\alias{flagInversions}
\title{Flag inversions}
\usage{
flagInversions(gr_ob, tol = Inf)
}
\arguments{
\item{gr_ob}{A \code{GenomicBreaks} object.}

\item{tol}{Tolerance window for the distance between two ranges.}
}
\value{
Returns the \code{GenomicBreaks} object with an extra \code{inv} metadata
column.
}
\description{
Flag ranges that start a triplet that would be colinear if the central pair
were aligned to the opposite strand.
}
\details{
Here is a trivial example of an inversion.\preformatted{┌───────────────┬───────────────┬───────────────┐
│ XSR:101-200:+ │ XSR:201-300:- │ XSR:301-400:+ │  (OKI2018_I69)
└───────────────┴───────────────┴───────────────┘
┌───────────────┬───────────────┬───────────────┐
│  S1:101-200:+ │  S1:201-300:+ │  S1:301-400:+ │  (OSKA2016)
└───────────────┴───────────────┴───────────────┘
}
}
\examples{
inv       <- GRanges(c("XSR:101-200:+", "XSR:201-300:-",  "XSR:301-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
flagInversions(inv)

}
\seealso{
Other Flagging functions: 
\code{\link{flagAll}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagInsersions}()},
\code{\link{flagInterruptions}()},
\code{\link{flagLongShort}()}

Other Inversion functions: 
\code{\link{filterInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{showInversions}()}
}
\concept{Flagging functions}
\concept{Inversion functions}
