% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagLongShort.R
\name{flagLongShort}
\alias{flagLongShort}
\title{Flag long and short arms}
\usage{
flagLongShort(gr, annot, select = c("first", "last", "arbitrary"))
}
\arguments{
\item{gr}{A \code{GBreaks} or a simple \code{GRanges} object}

\item{annot}{A \code{GRanges} file containing the coordinate of arms and their
nature (such as \code{long}, \code{short}, \code{XSR} or \code{YSR}) in a \code{Type}
\emph{metadata column}.}

\item{method}{One method name among \code{ignore}, \code{remove} and \code{split} to decide
on how to handle the regions that match both arms.}
}
\value{
Returns a modified version of the object with an extra \emph{metadata
column}, \code{Arm} in which the information from the annotation file's \code{Type}
column was transferred.  With \code{method = "ignore"}, ranges matching two arms
get \code{NA} as a value.  With \code{"remove"} they are deleted and with \code{"split"}
they are split arbitrarily.
}
\description{
\emph{Oikopleura} genomes are special as the long and short arms of their
chromosomes have different properties such as \code{GC} or repeat content.  It
can be useful to know if genomic region belongs to long or a short arm.
}
\examples{
gb       <- GRanges(c("chr1:101-180:+", "chr1:201-300:+",
                      "chr1:320-400:+", "chr1:501-550:+"))
gb$query <- GRanges(c("chrI:101-200",   "chrI:201-300",
                      "chrI:301-400",   "chrI:501-550"))
gb <- GBreaks(gb)
gb

annot <- GRanges(c("chr1:1-250", "chr1:260-500"))
annot$Type <- c("short", "long")
annot

flagLongShort(gb, annot)
flagLongShort(gb, annot, select = "last")
flagLongShort(gb, annot, select = "arbitrary")


}
\seealso{
Other Flagging functions: 
\code{\link{flagAll}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagInsersions}()},
\code{\link{flagInterruptions}()},
\code{\link{flagInversions}()}
}
\author{
Charles Plessy
}
\concept{Flagging functions}
