% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{flagTranslocations}
\alias{flagTranslocations}
\title{Flag translocations}
\usage{
flagTranslocations(gb, tol = Inf)
}
\arguments{
\item{gb}{A \code{GBreaks()} object.}

\item{tol}{Tolerance window for the distance between two ranges.}
}
\value{
Returns the \code{GBreaks} object with an extra \code{tra} metadata column.
}
\description{
Flag ranges that start a triplet that would be colinear if the central pair
were colinear, and that are an inversion.
}
\details{
Here is a trivial example of a translocation\preformatted{┌──────────────┬──────────────┬──────────────┐
│ chrA:101-200 │ chrA:201-300 │ chrA:301-400 │ (Target genome)
└──────────────┴──────────────┴──────────────┘
       +               +             +         (Alignment direction)
┌──────────────┬──────────────┬──────────────┐
│ chrB:101-200 │ chrC:201-300 │ chrB:301-400 │ (Query genome)
└──────────────┴──────────────┴──────────────┘
}
}
\examples{
flagTranslocations(exampleTranslocation)
flagTranslocations(exampleTranslocation2)
plotApairOfChrs(exampleTranslocation)
flagTranslocations(exampleDeletion)
flagTranslocations(exampleInsertion)
flagTranslocations(exampleInsertion |> swap() |> sort(ignore.strand = TRUE))
flagTranslocations(sort(reverse(exampleDeletion)))
flagTranslocations(exampleInversion)
flagTranslocations(exampleColinear3)

}
\seealso{
Other Flagging functions: 
\code{\link{flagAll}()},
\code{\link{flagColinearAlignments}()},
\code{\link{flagInversions}()},
\code{\link{flagLongShort}()}

Other Translocation functions: 
\code{\link{filterTranslocations}()},
\code{\link{showTranslocations}()}

Other Structural variants: 
\code{\link{StructuralVariants}},
\code{\link{flagInversions}()},
\code{\link{plotApairOfChrs}()}
}
\concept{Flagging functions}
\concept{Structural variants}
\concept{Translocation functions}
