% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{flipInversions}
\alias{flipInversions}
\title{Flip strand of inversions}
\usage{
flipInversions(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object processed with \code{\link{flagInversions}}.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are the
central part of an inversion triplet have their strand orientation flipped.
The \code{inv} metadata column is then discarded as it is not valid anymore.  As
the former inversion triplets are now collinear, new inversions may possibly
found after the object is coalesced again.
}
\description{
Flip strand of inversions
}
\examples{
exampleNestedInversions |> flagInversions()
flipInversions(exampleNestedInversions |> flagInversions())

}
\seealso{
Other Inversion functions: 
\code{\link{filterDoubleInversions}()},
\code{\link{filterInversions}()},
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{leftInversionGaps}()},
\code{\link{showInversions}()}

Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{forceSeqLengths}()},
\code{\link{guessSeqLengths}()},
\code{\link{keepLongestPair}()},
\code{\link{mergeSeqLevels}()},
\code{\link{reverse}()},
\code{\link{swap}()}
}
\concept{Inversion functions}
\concept{modifier functions}
