% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandNames.R
\name{flipStrandNames.default}
\alias{flipStrandNames.default}
\title{Reverse order and flip strand}
\usage{
\method{flipStrandNames}{default}(sn)
}
\value{
Returns the same class as its input.
}
\description{
Takes a character vector of "strand names", reverses the order of its
elements and flips the strand.
}
