% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapProportion.R
\name{gapProportion}
\alias{gapProportion}
\title{Proportion of gaps in alignment}
\usage{
gapProportion(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
A numeric value between 0 and 1.
}
\description{
The proportion of nucleotides that are part of an alignment but not matched
to another nucleotide.  If you divide it by the percent difference, you obtain
the S/I ratio studied by Chen and coll., 2009.
}
\note{
This is not a measure of distance. At greater evolutionary divergence,
genome alignment becomes increasingly difficult, leaving only the most
conserved regions alignable. Consequently, the proportion of gaps tends to
decrease.
}
\examples{
gapProportion(exampleSubstitutionMatrix)

}
\references{
Chen JQ, Wu Y, Yang H, Bergelson J, Kreitman M, Tian D. (2009).
Variation in the ratio of nucleotide substitution and indel rates across
genomes in mammals and bacteria. \emph{Mol Biol Evol.} 26(7):1523-31.
\doi{10.1093/molbev/msp063}
}
\seealso{
Other Alignment statistics:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GCequilibrium]{GCequilibrium()}},
\code{\link[=GCpressure]{GCpressure()}},
\code{\link[=GCproportion]{GCproportion()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link{exampleSubstitutionMatrix}},
\code{\link[=logDet_distance]{logDet_distance()}}
}
\author{
Charles Plessy
}
\concept{Alignment statistics}
