% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapProportion.R
\name{gapProportion}
\alias{gapProportion}
\title{Proportion of gaps in alignment}
\usage{
gapProportion(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
A numeric value between 0 and 1.
}
\description{
The proportion of nucleotides that are part of an alignment but not matched
to another nucleotide.  If you divide it by the percent difference, you obtain
the S/I ratio studied by Chen and coll., 2009.
}
\note{
This is not a measure of distance. At greater evolutionary divergence,
genome alignment becomes increasingly difficult, leaving only the most
conserved regions alignable. Consequently, the proportion of gaps tends to
decrease.
}
\examples{
parameters <- readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))
gapProportion(parameters)

}
\references{
Chen JQ, Wu Y, Yang H, Bergelson J, Kreitman M, Tian D. (2009).
Variation in the ratio of nucleotide substitution and indel rates across
genomes in mammals and bacteria. \emph{Mol Biol Evol.} 26(7):1523-31.
\doi{10.1093/molbev/msp063}
}
\seealso{
Other Nucleotide distances: 
\code{\link{F81_distance}()},
\code{\link{GCequilibrium}()},
\code{\link{GCpressure}()},
\code{\link{GCproportion}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{P_distance}()},
\code{\link{T92_distance}()}
}
\author{
Charles Plessy
}
\concept{Nucleotide distances}
