% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gb2comp}
\alias{gb2comp}
\title{Convert \code{GBreaks} object to a \code{comparison} object}
\usage{
gb2comp(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a \code{\link[genoPlotR:comparison]{genoPlotR::comparison}} object.
}
\description{
Convert \code{GBreaks} object to a \code{comparison} object
}
\note{
Note that the \code{comparison} objects assume that the ranges are on the
same sequence feature.  It is your responsibility that the input \code{GBreaks}
object also conforms to this assumption.
}
\examples{
inv       <- GRanges(c("XSR:101-180:+", "XSR:201-300:-",  "XSR:320-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
inv <- GBreaks(inv)
gb2comp(inv)

}
\author{
Charles Plessy
}
