% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gb2dna_seg}
\alias{gb2dna_seg}
\title{Convert \code{GBreaks} object to a list of \code{dna_seg} objects}
\usage{
gb2dna_seg(gb, ...)
}
\arguments{
\item{gb}{A \code{\link{GRanges}} or a \code{\link{GBreaks}} object.}

\item{...}{Extra arguments passed to \code{dna_seg()}.}
}
\value{
Returns a \code{list} of two \code{\link[genoPlotR:dna_seg]{genoPlotR::dna_seg}} objects, respectively
for the \emph{target} ranges the \emph{query} ranges.
}
\description{
Convert \code{GBreaks} object to a list of \code{dna_seg} objects
}
\examples{
inv       <- GRanges(c("XSR:101-180:+", "XSR:201-300:-",  "XSR:320-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
inv <- GBreaks(inv)
gb2dna_seg(inv)

}
\author{
Charles Plessy
}
