% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gb2dna_seg}
\alias{gb2dna_seg}
\title{Convert \code{GBreaks} object to a list of \code{dna_seg} objects}
\usage{
gb2dna_seg(gb, ...)
}
\arguments{
\item{gb}{A \code{\link{GRanges}} or a \code{\link{GBreaks}} object.}

\item{...}{Extra arguments passed to \code{dna_seg()}.}
}
\value{
Returns a \code{list} of two \code{\link[genoPlotR:dna_seg]{genoPlotR::dna_seg}} objects, respectively
for the \emph{target} ranges the \emph{query} ranges.
}
\description{
Convert \code{GBreaks} object to a list of \code{dna_seg} objects
}
\examples{
gb2dna_seg(exampleInversion)

}
\seealso{
Other genoPlotR functions: 
\code{\link{gb2comp}()},
\code{\link{gr2dna_seg}()},
\code{\link{plotApairOfChrs}()}

Other converter functions: 
\code{\link{gb2comp}()},
\code{\link{gb2xlim}()},
\code{\link{gr2dna_seg}()}
}
\author{
Charles Plessy
}
\concept{converter functions}
\concept{genoPlotR functions}
