% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicBreaks.R
\docType{class}
\name{GBreaks-class}
\alias{GBreaks-class}
\alias{GBreaks}
\title{The GenomicBreaks class}
\description{
The \code{GBreaks} class is a simple wrapper to the \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} class.
}
\details{
Aligned sequences of the \emph{target} genome are represented as the main
ranges of the \code{GRanges} object, and their counterparts in the \emph{query} genome
are represented as a \code{GRanges} object sored in the the metadata column
\code{query}.
}
\examples{
gb <- load_genomic_breaks(system.file("extdata/SacCer3__SacPar.gff3.gz", package = "GenomicBreaks"))
gb
gb$query

}
\seealso{
Other Bioconductor API functions: 
\code{\link{getSeq}()},
\code{\link{pairwiseAlignment}()},
\code{\link{range_GBreaks}()},
\code{\link{reverse}()},
\code{\link{subsetByOverlaps_GBreaks}()}
}
\concept{Bioconductor API functions}
