% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCpressure.R
\name{GCpressure}
\alias{GCpressure}
\title{GC pressure}
\usage{
GCpressure(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
A numeric value between \eqn{-1} and \eqn{1}.
}
\description{
Computes the following equation:
\deqn{\frac{\text{W}\to\text{S} \quad-\quad \text{S}\to\text{W}}{\text{W}\to\text{S} \quad+\quad \text{S}\to\text{W}}}{GC* = (W→S - S→W) / (W→S + S→W)}
}
\note{
This index was suggested to by by M365 Copilot while using it to write
the \code{\link{GCequilibrium}} function and its documentation.  Despite my attempts I
could not find a paper that uses it explicitely…  I am including it in this
package because I am screening as many indices that I can extract from an
alignment pair.
}
\examples{
GCpressure(exampleSubstitutionMatrix)
GCpressure(t(exampleSubstitutionMatrix))

}
\seealso{
Other Alignment statistics: 
\code{\link{F81_distance}()},
\code{\link{GCequilibrium}()},
\code{\link{GCproportion}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{P_distance}()},
\code{\link{T92_distance}()},
\code{\link{exampleSubstitutionMatrix}},
\code{\link{gapProportion}()}
}
\author{
Charles Plessy
}
\concept{Alignment statistics}
