% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCproportion.R
\name{GCproportion}
\alias{GCproportion}
\title{Proportion of GC in alignment}
\usage{
GCproportion(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
A numeric value between 0 and 1.
}
\description{
The proportion of GC in the \emph{target} genome in an alignment.  For the same
on the \emph{query} genome, just transpose the input matrix.
}
\examples{
GCproportion(exampleSubstitutionMatrix)
GCproportion(t(exampleSubstitutionMatrix))

}
\seealso{
Other Alignment statistics: 
\code{\link{F81_distance}()},
\code{\link{GCequilibrium}()},
\code{\link{GCpressure}()},
\code{\link{HKY85_distance}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{P_distance}()},
\code{\link{T92_distance}()},
\code{\link{TN93_distance}()},
\code{\link{exampleSubstitutionMatrix}},
\code{\link{gapProportion}()},
\code{\link{logDet_distance}()}
}
\author{
Charles Plessy
}
\concept{Alignment statistics}
