% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bps.R
\name{get_bps}
\alias{get_bps}
\title{Breakpoints}
\usage{
get_bps(gr_ob)
}
\arguments{
\item{gr_ob}{GRanges object containing pairwise alignment}
}
\value{
GRanges object of the breakpoints
}
\description{
Given a GRanges object, the function produces a GRanges object detailing the breakpoints only. The only converted data is that of the main GRanges subject, and no metadata is processed or carried through
}
\examples{
gr <- GRanges(c("chr2", "chr2", "chr1", "chr3"), IRanges(1:4, width=4:1))
get_bps(gr)
}
