% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeq.R
\name{getSeq}
\alias{getSeq}
\alias{getSeq,GRanges-method}
\alias{getSeq,GBreaks-method}
\title{Get genomic sequences}
\usage{
getSeq(x, ...)

\S4method{getSeq}{GRanges}(x, ...)

\S4method{getSeq}{GBreaks}(x, ...)
}
\arguments{
\item{x}{A \code{GBreaks} or a \code{GRanges} object.}

\item{...}{Extra arguments (not used, but present for compatibility).}
}
\value{
A \code{\link[Biostrings:XStringSet-class]{Biostrings::DNAStringSet}} object containing the extracted
sequence(s).
}
\description{
Extract the sequences of the genomic ranges from a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}}
or a \code{\link{GBreaks}} object.
}
\examples{
Scerevisiae <- BSgenome.Scerevisiae.UCSC.sacCer3::Scerevisiae
getSeq(GRanges("chrI:1-20", seqinfo = seqinfo(Scerevisiae)))

}
\seealso{
See also the \code{\link[Biostrings:getSeq]{Biostrings::getSeq}} function.

Other Bioconductor API functions: 
\code{\link{GBreaks-class}},
\code{\link{pairwiseAlignment}()},
\code{\link{range_GBreaks}()},
\code{\link{reverse}()},
\code{\link{subsetByOverlaps_GBreaks}()}
}
\author{
Charles Plessy
}
\concept{Bioconductor API functions}
