% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOC.R
\name{GOC}
\alias{GOC}
\title{Gene Order Conservation}
\usage{
GOC(gb, vicinity = 4, debug = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{vicinity}{How far to search for the neighbour orthologue.}

\item{debug}{See below.}
}
\value{
Returns a numeric value between 0 and 1.  If \code{debug = TRUE},
returns a copy of the \code{gb} object with additional columns showing details of
the computation.
}
\description{
The Gene Order Conservation (GOC) number defined by Rocha (2003) is:
\dQuote{\emph{the average number of orthologues for which the consecutive
orthologue co-occurs close by in the other genome. It varies between 0
(no co-occurrence) and 1 (complete gene order conservation)}}.
}
\note{
Note that calculating GOC on whole-genome alignments is not expected
to produce meaningful results.  This function is more useful when comparing
the position of orthologues, represented in a \code{GBreaks} object.
}
\examples{
exampleColinear
GOC(exampleColinear)

exampleTranslocation
GOC(exampleTranslocation)
GOC(exampleTranslocation, v=1)

# GOC computation is strandless
exampleInversion
GOC(exampleInversion)

}
\references{
Rocha, Eduardo P C. \dQuote{DNA repeats lead to the accelerated
loss of gene order in bacteria.} \emph{Trends in genetics : TIG} vol. 19,11
(2003): 600-3. \href{https://doi.org/10.1016/j.tig.2003.09.011}{doi:10.1016/j.tig.2003.09.011}
}
\seealso{
Other Colinearity functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=chain_contigs]{chain_contigs()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=dist2next]{dist2next()}},
\code{\link[=filterColinearRegions]{filterColinearRegions()}},
\code{\link[=flagColinearAlignments]{flagColinearAlignments()}},
\code{\link[=flagPairs]{flagPairs()}}

Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\concept{Colinearity functions}
\concept{Similarity indexes}
