% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOC.R
\name{GOC}
\alias{GOC}
\title{Gene Order Conservation}
\usage{
GOC(gb, vicinity = 4, debug = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{vicinity}{How far to search for the neighbour orthologue.}
}
\value{
Returns a numeric value between 0 and 1.  If \code{debug = TRUE},
returns a copy of the \code{gb} object with additional columns showing details of
the computation.
}
\description{
The Gene Order Conservation (GOC) number defined by Rocha (2003) is:
\dQuote{\emph{the average number of orthologues for which the consecutive
orthologue co-occurs close by in the other genome. It varies between 0
(no co-occurrence) and 1 (complete gene order conservation)}}.
}
\examples{
gb1       <- GRanges(c(A="Ref:100-200:+", B="Ref:400-500:+"))
gb1$query <- GRanges(c(A="Que:100-200",   B="Que:400-500"))
gb1
GOC(gb1)

gb4       <- GRanges(c("Ref:100-200:+",   "Ref:400-500:+",   "Ref:600-700:+"))
gb4$query <- GRanges(c("Que:1100-1200:+", "Que:1700-1800:+", "Que:1500-1600:+"))
gb4
GOC(gb4)
GOC(gb4, vicinity = 1)

inv       <- GRanges(c("XSR:101-180:+", "XSR:201-300:-",  "XSR:320-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
inv
GOC(inv, vicinity = 1)

}
\references{
Rocha, Eduardo P C. \dQuote{DNA repeats lead to the accelerated
loss of gene order in bacteria.} \emph{Trends in genetics : TIG} vol. 19,11
(2003): 600-3. \href{https://doi.org/10.1016/j.tig.2003.09.011}{doi:10.1016/j.tig.2003.09.011}
}
\seealso{
Other Colinearity functions: 
\code{\link{chain_contigs}()},
\code{\link{coalesce_contigs}()},
\code{\link{flagColinearAlignments}()}

Other Similarity indexes: 
\code{\link{correlation_index}()},
\code{\link{synteny_index}()}
}
\concept{Colinearity functions}
\concept{Similarity indexes}
