% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gr2dna_seg}
\alias{gr2dna_seg}
\title{Convert \code{GRanges} object to \code{dna_seg} format}
\usage{
gr2dna_seg(gr, ...)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{...}{Extra arguments passed to \code{dna_seg()}.}
}
\description{
The \code{\link[genoPlotR:dna_seg]{genoPlotR::dna_seg}} class represents \emph{DNA segments} using
a \code{data.frame} format.
}
\examples{
inv       <- GRanges(c("XSR:101-180:+", "XSR:201-300:-",  "XSR:320-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
gr2dna_seg(inv)
gr2dna_seg(inv$query)

}
\seealso{
Other genoPlotR functions: 
\code{\link{gb2comp}()},
\code{\link{gb2dna_seg}()},
\code{\link{plotApairOfChrs}()}

Other converter functions: 
\code{\link{gb2comp}()},
\code{\link{gb2dna_seg}()}
}
\author{
Charles Plessy
}
\concept{converter functions}
\concept{genoPlotR functions}
