% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{gr2dna_seg}
\alias{gr2dna_seg}
\title{Convert \code{GRanges} object to \code{dna_seg} format}
\usage{
gr2dna_seg(gr, ...)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{...}{Extra arguments passed to \code{dna_seg()}.}
}
\value{
Returns a \code{genoPlotR::dna_seq} or \code{NULL} if the input object has a
length of zero.
}
\description{
The \code{\link[genoPlotR:dna_seg]{genoPlotR::dna_seg}} class represents \emph{DNA segments} using
a \code{data.frame} format.
}
\examples{
gr2dna_seg(exampleInversion)
gr2dna_seg(exampleInversion$query)

}
\seealso{
Other genoPlotR functions: 
\code{\link{gb2comp}()},
\code{\link{gb2dna_seg}()},
\code{\link{plotApairOfChrs}()}

Other converter functions: 
\code{\link{gb2comp}()},
\code{\link{gb2dna_seg}()},
\code{\link{gb2xlim}()}
}
\author{
Charles Plessy
}
\concept{converter functions}
\concept{genoPlotR functions}
