% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessSeqLenghts.R
\name{guessSeqLengths}
\alias{guessSeqLengths}
\title{Guesstimate seqlevel lengths}
\usage{
guessSeqLengths(gr)
}
\arguments{
\item{gr}{A \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges}} object}
}
\value{
Returns sequence lengths that have been guessed as the maximal end
coordinates found in the \code{gr}, or the \code{gr}'s \code{seqlengths} if if they did
already exist.
}
\description{
When no \code{\link[Seqinfo:seqlengths]{Seqinfo::seqlengths}} are available, one can resort to set
them as to the maximal end coordinate found in the object.
}
\examples{
exampleTranslocation
seqlengths(exampleTranslocation$query)
guessSeqLengths(exampleTranslocation$query)
gb2 <- exampleTranslocation
seqlengths(gb2$query) <- NA
guessSeqLengths(gb2$query)

}
\seealso{
Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\author{
Charles Plessy
}
\concept{modifier functions}
