% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessSeqLenghts.R
\name{guessSeqLengths}
\alias{guessSeqLengths}
\title{Guesstimate seqlevel lengths}
\usage{
guessSeqLengths(gr)
}
\arguments{
\item{gr}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object}
}
\value{
Returns sequence lengths that have been guessed as the maximal end
coordinates found in the \code{gr}, or the \code{gr}'s \code{seqlengths} if if they did
already exist.
}
\description{
When no \code{\link[GenomeInfoDb:seqinfo]{GenomeInfoDb::seqlengths}} are available, one can resort to set
them as to the maximal end coordinate found in the object.
}
\examples{
exampleTranslocation
seqlengths(exampleTranslocation$query)
guessSeqLengths(exampleTranslocation$query)
gb2 <- exampleTranslocation
seqlengths(gb2$query) <- NA
guessSeqLengths(gb2$query)

}
\seealso{
Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{flipInversions}()},
\code{\link{forceSeqLengths}()},
\code{\link{keepLongestPair}()},
\code{\link{mergeSeqLevels}()},
\code{\link{reverse}()},
\code{\link{splitSeqLevel}()},
\code{\link{swap}()}
}
\author{
Charles Plessy
}
\concept{modifier functions}
