% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessSeqLenghts.R
\name{guessSeqLengths}
\alias{guessSeqLengths}
\title{Guesstimate seqlevel lengths}
\usage{
guessSeqLengths(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object}
}
\value{
\code{\link{guessSeqLengths}} returns sequence lengths that have been guessed
as the maximal end coordinates found in the object, or the object's
\code{seqlengths} if if they did already exist.
}
\description{
When no \code{\link{seqlengths}} are available, one can resort to set them as to the
maximal end coordinate found in the object.
}
\examples{
exampleTranslocation
seqlengths(exampleTranslocation$query)
guessSeqLengths(exampleTranslocation$query)
gb2 <- exampleTranslocation
seqlengths(gb2$query) <- NA
guessSeqLengths(gb2$query)

}
\seealso{
Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{forceSeqLengths}()},
\code{\link{mergeSeqLevels}()},
\code{\link{reverse}()},
\code{\link{swap}()}
}
\author{
Charles Plessy
}
\concept{modifier functions}
