% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversionDistance.R
\name{inversionDistance}
\alias{inversionDistance}
\title{Inversion Distance}
\usage{
inversionDistance(gb_object)
}
\arguments{
\item{gb_object}{A GBreaks object containing genome data from which a permutation vector can be extracted
using the \code{permutationVector()} function.}
}
\value{
An integer: the minimal number of inversions needed to sort the permutation.
}
\description{
Computes the minimal number of inversions required to sort a signed permutation
using the Hannenhalli and Pevzner algorithm.
}
\details{
This function uses several internal helper functions (e.g., \code{extendedPermutation},
\code{breakpoint_graph}, \code{hurdles_count}, \code{superhurdles_count}, and others) to
compute properties of the breakpoint graph and identify cycles, hurdles and superhurdles.
It also depends on \code{permutationVector()}, which is defined in another .R file.

This algorithm was designed to work in \emph{a single chromosome alignment}.
Although the function still works if the GBreaks object involves more than one chromosome,
the returned value for the minimal number of inversions will imply in non-usual inversions
if different chromosomes have orthologous regions.
}
\seealso{
\code{\link{permutationVector}} for generating the permutation vector.
}
\author{
Bruna Fistarol
}
