% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversionEstimate_BD.R
\name{inversionEstimate_BD_many}
\alias{inversionEstimate_BD_many}
\title{This function should be preferred when there are many inversion estimates to be computed.}
\usage{
inversionEstimate_BD_many(n, obs_nb_cycles, cyc_all = NA)
}
\arguments{
\item{n}{An integer: The number of markers.
A marker represents an aligned block, a gene,
or any other genomic region being rearranged.}

\item{obs_nb_cycles}{An integer: The observed number of cycles
in the breakpoint graph of two genomes.}
}
\description{
This function should be preferred when there are many inversion estimates to be computed.
}
\author{
Priscila Biller
}
