% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{is_fortress}
\alias{is_fortress}
\title{Check fortrees}
\usage{
is_fortress(superhurdles)
}
\arguments{
\item{superhurdles}{Object that stores the counts of hurdles and superhurdles in the breakpoint graph.}
}
\description{
Check whether or not the permutation is a fortress.
A permutation is a fortress if the number of hurdles is odd and all hurdles are superhurdles.
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=cycle_nontrivial_count]{cycle_nontrivial_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_interleaving]{is_interleaving()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
