% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointGraph.R
\name{is_interleaving}
\alias{is_interleaving}
\title{Interleaving Check}
\usage{
is_interleaving(p_extended, graph_1, graph_2, edge_1, edge_2)
}
\description{
Interleaving Check
}
\seealso{
Other Breakpoint graph functions:
\code{\link[=bp_count]{bp_count()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=breakpoint_graph]{breakpoint_graph()}},
\code{\link[=components_graph]{components_graph()}},
\code{\link[=cycle_count]{cycle_count()}},
\code{\link[=extendedPermutation]{extendedPermutation()}},
\code{\link[=hurdles_count]{hurdles_count()}},
\code{\link[=is_fortress]{is_fortress()}},
\code{\link[=superhurdles_count]{superhurdles_count()}}
}
\author{
Bruna Fistarol
}
\concept{Breakpoint graph functions}
