% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JC69_distance.R
\name{JC69_distance}
\alias{JC69_distance}
\title{Jukes-Cantor 1969 distance}
\usage{
JC69_distance(m, tot = 1, adjust_p = FALSE)
}
\arguments{
\item{m}{Two possible types of input are available:
\enumerate{
\item A matrix of counts or probabilities for bases of the \emph{target} genome to be aligned to bases on the \emph{query} genome;
\item The mismatch count (a positive integer) or proportion (a number between \code{0.0} and \code{1.0}).
}}

\item{tot}{The total count of nucleotides, in case the parameter \code{m} is the \emph{count} of mismatches.}

\item{adjust_p}{A boolean flag. If \code{TRUE}, the distance is scaled between \code{0} and \code{0.75} to ensure the logarithm stays positive.}
}
\value{
Returns a numeric value representing the evolutionary distance between two genomes. The greater the value, the more genetically different the genomes are.
}
\description{
The Jukes and Cantor 1969 (JC69) distance is calculated as \eqn{-\frac{3}{4} \ln(1 - \frac{4}{3} p )}, where \eqn{p} is the sum of the probabilities of the alignment. See the \href{https://en.wikipedia.org/wiki/Models_of_DNA_evolution#JC69_model_(Jukes_and_Cantor_1969)}{Wikipedia} for more details.
}
\examples{
# Using a substitution matrix as input:
d <- JC69_distance(exampleSubstitutionMatrix)

# Using numerical values as input:
d <- JC69_distance(25, 100)

}
\references{
Jukes, T.H. & Cantor, C.R. (1969). "Evolution of protein molecules." In \emph{Mammalian Protein Metabolism} (pp. 21–132). Academic Press.
}
\seealso{
Other Alignment statistics:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GCequilibrium]{GCequilibrium()}},
\code{\link[=GCpressure]{GCpressure()}},
\code{\link[=GCproportion]{GCproportion()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link{exampleSubstitutionMatrix}},
\code{\link[=gapProportion]{gapProportion()}},
\code{\link[=logDet_distance]{logDet_distance()}}

Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Zikun Yang

Priscila Biller
}
\concept{Alignment statistics}
\concept{Similarity indexes}
