% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JC69_distance.R
\name{JC69_distance}
\alias{JC69_distance}
\title{Index representing the evolutionary distance between two genomes.}
\usage{
JC69_distance(train_parameters)
}
\arguments{
\item{train_parameters}{A list containing the probabilities of the alignment.}
}
\value{
Returns a numeric value show the evolutionary distance between two genomes. the larger the value, the more different the two genomes are.
}
\description{
The JC69 distance is calculated as -0.75 * log(1 - 4 * p / 3), where p is the sum of the probabilities of the alignment. See \href{https://en.wikipedia.org/wiki/Models_of_DNA_evolution#JC69_model_(Jukes_and_Cantor_1969)}{WIKI} for more details.
}
\note{
Here, the probability ranges from 0 to 1. The list input can contain other information, but the function will only use the name in this format probability_{REF}_{ALT}
}
\examples{
parameters <- readTrainFile("train.txt")
JC69_distance(parameters)

}
\seealso{
Other Similarity indexes: 
\code{\link{GOC}()},
\code{\link{K80_distance}()},
\code{\link{T92_distance}()},
\code{\link{correlation_index}()},
\code{\link{karyotype_change_index}()},
\code{\link{readTrainFile}()},
\code{\link{strand_randomisation_index}()},
\code{\link{synteny_index}()}
}
\concept{Similarity indexes}
