% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K80_distance.R
\name{K80_distance}
\alias{K80_distance}
\title{Kimura's 2-parameter distance}
\usage{
K80_distance(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
Returns a numeric value show the evolutionary distance between two genomes. the larger the value, the more different the two genomes are.
}
\description{
The Kimura's 2-parameter (K80) distance distinguishes between transitions and transversions.  It is calculated as \eqn{-\frac{1}{2} \ln\left(\left(1 - 2p - q\right) \times \sqrt{1 - 2q}\right)}, where \eqn{p} is the probability of transition and \eqn{q} is the probability of transversion. See the \href{https://en.wikipedia.org/wiki/Models_of_DNA_evolution#K80_model_(Kimura_1980)}{Wikipedia} for more details.
}
\examples{
parameters <- readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))
K80_distance(parameters)

}
\references{
Kimura, M. (1980). "A simple method for estimating evolutionary rates of base substitutions through comparative studies of nucleotide sequences." \emph{Journal of Molecular Evolution}, 16, 111–120. DOI: \href{https://doi.org/10.1007/BF01731581}{10.1007/BF01731581}
}
\seealso{
Other Nucleotide distances: 
\code{\link{F81_distance}()},
\code{\link{JC69_distance}()},
\code{\link{P_distance}()},
\code{\link{T92_distance}()},
\code{\link{gapProportion}()}
}
\author{
Zikun Yang
}
\concept{Nucleotide distances}
