% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K80_distance.R
\name{K80_distance}
\alias{K80_distance}
\title{Index representing the evolutionary distance between two genomes.}
\usage{
K80_distance(train_parameters)
}
\arguments{
\item{train_parameters}{A list containing the probabilities of the alignment.}
}
\value{
Returns a numeric value show the evolutionary distance between two genomes. the larger the value, the more different the two genomes are.ss
}
\description{
The K80 distance is calculated as -0.5 * log((1 - 2 * p - q) * sqrt(1 - 2 * q)). See \href{https://en.wikipedia.org/wiki/Models_of_DNA_evolution#K80_model_(Kimura_1980)}{WIKI} for more details.
}
\note{
Here, the probability ranges from 0 to 1. The list input can contain other information, but the function will only use the name in this format probability_{REF}_{ALT}
}
\examples{
parameters <- readTrainFile("train.txt")
K80_distance(parameters)

}
\seealso{
Other Similarity indexes: 
\code{\link{GOC}()},
\code{\link{JC69_distance}()},
\code{\link{T92_distance}()},
\code{\link{correlation_index}()},
\code{\link{karyotype_change_index}()},
\code{\link{readTrainFile}()},
\code{\link{strand_randomisation_index}()},
\code{\link{synteny_index}()}
}
\concept{Similarity indexes}
