% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K80_gap_distance.R
\name{K80_gap_distance}
\alias{K80_gap_distance}
\title{Kimura's 2-parameter distance extended with gaps}
\usage{
K80_gap_distance(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome. As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
Returns a numeric value showing the evolutionary distance between
two genomes. The larger the value, the more different the two genomes are.
}
\description{
Nishimaki and Sato's K2P + Gap distance extends Kimura's 2-parameter
distance by treating gaps as insertion/deletion states instead of removing
gap-containing sites.
}
\details{
The distance is calculated as
\deqn{
K = \frac{3}{4} w \log(w)
    - \frac{w}{2} \log\left((S - P) \sqrt{S + P - Q}\right)
}
where \eqn{S} is the probability of identical nucleotide pairs,
\eqn{P} is the probability of transition-type nucleotide pairs,
\eqn{Q} is the probability of transversion-type nucleotide pairs, and
\eqn{w} is the nucleotide occupancy probability.

When there are no gaps, \eqn{w = 1}, and this expression reduces to the
usual K80 distance.
}
\examples{
K80_gap_distance(exampleSubstitutionMatrix)

# When there are no gaps, it returns the same as the K80 distance
nogaps <- exampleSubstitutionMatrix
nogaps["-",] <- 0
nogaps[,"-"] <- 0
K80_gap_distance(nogaps)
K80_distance(nogaps)

}
\references{
Kimura, M. (1980). "A simple method for estimating evolutionary rates of base
substitutions through comparative studies of nucleotide sequences."
\emph{Journal of Molecular Evolution}, 16, 111–120.
DOI: \doi{10.1007/BF01731581}

Nishimaki, T. and Sato, K. (2019). "An Extension of the Kimura
Two-Parameter Model to the Natural Evolutionary Process."
\emph{Journal of Molecular Evolution}, 87, 60–67.
DOI: \doi{10.1007/s00239-018-9885-1}
}
\seealso{
Other Alignment statistics:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GCequilibrium]{GCequilibrium()}},
\code{\link[=GCpressure]{GCpressure()}},
\code{\link[=GCproportion]{GCproportion()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=JC69_distance_allseq]{JC69_distance_allseq()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link{exampleSubstitutionMatrix}},
\code{\link[=gapProportion]{gapProportion()}},
\code{\link[=logDet_distance]{logDet_distance()}}

Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=JC69_distance_allseq]{JC69_distance_allseq()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Charles Plessy
}
\concept{Alignment statistics}
\concept{Similarity indexes}
