% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype_index.R
\name{karyotype_index}
\alias{karyotype_index}
\title{Index representing how the karyotype changes between chromosomes of two genomes}
\usage{
karyotype_index(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a numeric value between 0 and 1. 0 is expected for same karyotype genomes.
}
\description{
The index is calculated as the mean of the number of chromosomes in the two genomes.  ###############
}
\note{
Here, the definition of \sQuote{karyotype change} is 1 - (the mean value of change score for each chromosome in the reference genome.
The change score for chromosome Ref is calculated as 1 / (number of mapped chromosomes in the query genome). To be noticed,
the mapped length of mapped chromosome in the query genome should be at least 10\% of the length of the reference chromosome.
if chromosome Ref has only one mapped chromosome in the query genome, the change score is 1. For 2 chromosomes, the change score is 0.5.
For the symmetry, the sum of change scores for each chromosome in query will be normalized to 1 if the sum is more than 1.
}
\examples{
gb       <- GRanges(c("Ref:100-200:+",   "Ref:400-500:+",    "Re2:600-700:+"))
gb$query <- GRanges(c("Que:1100-1200:+", "Que2:1700-1800:+", "Que2:1500-1600:+"))
karyotype_index(gb)

}
\seealso{
Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\concept{Similarity indexes}
