% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{leftInversionGaps}
\alias{leftInversionGaps}
\title{Isolate the left-side gaps in inversions}
\usage{
leftInversionGaps(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a \code{\link{GRanges}} object representing the left-side gaps in the
\code{GBreaks} object.
}
\description{
Isolate the left-side gaps in inversions
}
\examples{
inv       <- GRanges(c("XSR:101-180:+", "XSR:201-300:-",  "XSR:320-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
inv <- flagInversions(inv)
leftInversionGaps(inv)

inv2       <- GRanges(c("XSR:101-180:-", "XSR:201-300:+",  "XSR:320-400:-"))
inv2$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
inv2 <- flagInversions(inv2)
leftInversionGaps(inv2)

}
\seealso{
Other Inversion functions: 
\code{\link{filterInversions}()},
\code{\link{flagInversions}()},
\code{\link{showInversions}()}
}
\concept{Inversion functions}
