% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genomic_breaks.R
\name{load_genomic_breaks}
\alias{load_genomic_breaks}
\title{Load pairwise genome alignments}
\usage{
load_genomic_breaks(file, bsgenome_ref, bsgenome_query)
}
\arguments{
\item{file}{Path to a file in GFF3 format}

\item{bsgenome_ref}{A \code{BSgenome} object representing the reference genome.}

\item{bsgenome_query}{A \code{BSgenome} object representing the query genome.}
}
\value{
Returns a \verb{GRanges object} where each element represents a pairwise
alignment block.  The \code{granges} part of the object contains the coordinates
on the \emph{reference} genome, and the \code{query} metadata column contains the query
coordinates in \code{GRanges} format.  The \code{seqinfo} of each \code{BSgenome} object
passed as parameters are copied to the \code{GRanges} objects accordingly.
}
\description{
Loads alignments of a query genome to a reference genome from a text file in
GFF3 format
}
\details{
This function expects the pairwise alignment to be represented in GFF3 format
in the following way:
\itemize{
\item Alignments blocks are represented by \code{match_part} entries in the \emph{type}
column  Other entries will be discarded.
\item The coordinate system of the file is the one of the \emph{reference} genome.
\item The \code{Target} tag in the \emph{attribute} column contains the coordinates of
the match in the \emph{query} genome.
\item Stand information is set so that \emph{query} genome coordinates are always on
the \emph{plus} strand.
}
}
