% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInversions.R
\name{makeInversion.chr}
\alias{makeInversion.chr}
\title{Make an inversion}
\usage{
makeInversion.chr(gr, dist = runif)
}
\arguments{
\item{gr}{\code{\link{GenomicRanges}} object}

\item{dist}{Distribution function for new breakpoints on \code{gr}'s sequence levels.}
}
\description{
Make an inversion
}
\examples{
gr <- GRanges("chr1:1-1000000:+")
gr2 <- GRanges(c("chr1:1-1000000:+", "chr2:1-1000000:+"))
gr3 <- GRanges(c("chr1:1-1000000:+", "chr1:1000000-2000000:+"))
gr4 <- GRanges(c("chr1:1-1000000:+", "chr1:1000000-2000000:-"))
gr5 <- GRanges(c("chr1:1-1000000:-"))
}
