% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOxfordPlots.R
\name{makeOxfordPlots}
\alias{makeOxfordPlots}
\title{Make Oxford Plots}
\usage{
makeOxfordPlots(
  gb,
  sp1Name = "sp1",
  sp2Name = "sp2",
  sp1ChrArms = NULL,
  sp2ChrArms = NULL,
  type = c("line", "point"),
  size = 1
)
}
\arguments{
\item{gb}{A \code{GBreaks} object}

\item{sp1Name}{Name of the first species (default: sp1)}

\item{sp2Name}{Name of the second species (default: sp2)}

\item{sp1ChrArms}{A \code{GBreaks} object of chromosome arms in sp1 genome}

\item{sp2ChrArms}{A \code{GBreaks} object of chromosome arms in sp2 genome}

\item{type}{The type of the plot (\code{point} or \code{line})}

\item{size}{The size of the plotted dots or segments.}
}
\value{
Returns a \code{ggplot2} object that can be further modified using the
\code{+} operatore.
}
\description{
Takes a \code{\link{GBreaks}} object and prepares an Oxford (macrosynteny) of the
coordinates of the \emph{query} ranges against the \emph{target} ranges after
concatenating them.
}
\examples{
makeOxfordPlots(exampleTranslocation)
makeOxfordPlots(exampleTranslocation, type = "p")

}
\seealso{
Other plot functions: 
\code{\link{plotApairOfChrs}()}
}
\author{
Aleksandra Bliznina
}
\concept{plot functions}
