% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOxfordPlots.R
\name{makeOxfordPlots}
\alias{makeOxfordPlots}
\title{Make Oxford Plots}
\usage{
makeOxfordPlots(
  gb,
  selChroms = NULL,
  sp1Name = "sp1",
  sp2Name = "sp2",
  sp1ChrArms = NULL,
  sp2ChrArms = NULL
)
}
\arguments{
\item{gb}{A GenomicBreaks object}

\item{selChroms}{A list of chromosomes in target genome to plot}

\item{sp1Name}{Name of the first species (default: sp1)}

\item{sp2Name}{Name of the second species (default: sp2)}

\item{sp1ChrArms}{A GenomicBreaks object of chromosome arms in sp1 genome}

\item{sp2ChrArms}{A Genomic Breaks object of chromosome arms in sp2 genome}
}
\value{
Returns an Oxford (macrosynteny) plot generated with \code{ggplot2}
}
\description{
TBD
}
\examples{
gb       <- GRanges(c("XSR:101-200:+", "XSR:201-300:-",  "XSR:301-400:+", "chrUn_1:401-500:+"))
gb$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400", "S1:401-500"))
makeOxfordPlots(gb)

}
\seealso{
Other plot functions: 
\code{\link{plotApairOfChrs}()}
}
\author{
Aleksandra Bliznina
}
\concept{plot functions}
