% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOxfordPlots.R
\name{makeOxfordPlots}
\alias{makeOxfordPlots}
\title{Make Oxford Plots}
\usage{
makeOxfordPlots(
  gb,
  sp1Name = "sp1",
  sp2Name = "sp2",
  sp1ChrArms = NULL,
  sp2ChrArms = NULL,
  type = c("line", "point"),
  size = 1
)
}
\arguments{
\item{gb}{A GenomicBreaks object}

\item{sp1Name}{Name of the first species (default: sp1)}

\item{sp2Name}{Name of the second species (default: sp2)}

\item{sp1ChrArms}{A GenomicBreaks object of chromosome arms in sp1 genome}

\item{sp2ChrArms}{A Genomic Breaks object of chromosome arms in sp2 genome}

\item{type}{The type of the plot (\code{point} or \code{line})}

\item{size}{The size of the plotted dots or segments.}
}
\value{
Returns an Oxford (macrosynteny) plot generated with \code{ggplot2}
}
\description{
TBD
}
\examples{
makeOxfordPlots(exampleTranslocation)
makeOxfordPlots(exampleTranslocation, type = "p")

}
\seealso{
Other plot functions: 
\code{\link{plotApairOfChrs}()}
}
\author{
Aleksandra Bliznina
}
\concept{plot functions}
