% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOxfordPlots.R
\name{makeOxfordPlots}
\alias{makeOxfordPlots}
\title{Make Oxford Plots}
\usage{
makeOxfordPlots(
  gb,
  sp1Name = "target",
  sp2Name = "query",
  sp1ChrArms = NULL,
  sp2ChrArms = NULL,
  type = c("line", "point", "none"),
  size = 1,
  diag = TRUE,
  col = c("seqnames", "strand", "score")
)
}
\arguments{
\item{gb}{A \code{GBreaks} object}

\item{sp1Name}{Name of the first species (default: sp1)}

\item{sp2Name}{Name of the second species (default: sp2)}

\item{sp1ChrArms}{A \code{GBreaks} object of chromosome arms in sp1 genome}

\item{sp2ChrArms}{A \code{GBreaks} object of chromosome arms in sp2 genome}

\item{type}{The type of the plot (\code{point}, \code{line} or \code{none})}

\item{size}{The size of the plotted dots or segments.}

\item{diag}{Diagonalise the plot by reordering \emph{query} sequence levels.}

\item{col}{Colour of the plotted dots or lines by \code{seqnames}, \code{strand} or
\code{score}.}
}
\value{
Returns a \code{ggplot2} object that can be further modified using the
\code{+} operator.  Use \code{type = 'none'} to receive an object without \emph{geom}
layer.
}
\description{
Takes a \code{\link{GBreaks}} object and prepares an Oxford (macrosynteny) of the
coordinates of the \emph{query} ranges against the \emph{target} ranges after
concatenating them.
}
\examples{
makeOxfordPlots(exampleTranslocation)
makeOxfordPlots(exampleTranslocation, type = "p")

}
\seealso{
Other plot functions: 
\code{\link{bp_heatmap}()},
\code{\link{bp_pair_analysis}()},
\code{\link{feature_coverage}()},
\code{\link{plotApairOfChrs}()}
}
\author{
Aleksandra Bliznina

Charles Plessy
}
\concept{plot functions}
