% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPairs.R
\name{matchPairs}
\alias{matchPairs}
\title{Match pairs of sequences}
\usage{
matchPairs(gb, drop = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{drop}{Drop unused sequence levels.}
}
\value{
A \code{GBreaks} object.
}
\description{
For each sequence of the \emph{target} genome, assign a sequence of the \emph{query}
genome, and discard all the ranges that are not from matched pairs.
}
\note{
This is not a reciprocal best match function, because it assumes that
the genomes are close enough and that there were no major karyotype changes,
so that the pairing is a trivial operation.
}
\examples{
matchPairs(exampleTranslocation)
}
\seealso{
Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=chromosome_mapping]{chromosome_mapping()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeDoubleInversions]{removeDoubleInversions()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\author{
Charles Plessy
}
\concept{modifier functions}
