% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_distance.R
\name{P_distance}
\alias{P_distance}
\title{Percent difference Distance}
\usage{
P_distance(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
A numeric value representing the evolutionary distance between two
genomes.  The larger the value, the more divergent the genomes.
}
\description{
The P-distance is simply the percentage of the aligned base pairs where the
two bases are different.  See May 2004 for a discussion on the many ways to
compute percent identity (and therefore difference.)
}
\examples{
parameters <- readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))
P_distance(parameters)

}
\references{
May, A. C. W. (2004). Percent Sequence Identity: The Need to Be Explicit.
\emph{Structure}, 12(5), 737–738. \doi{10.1016/j.str.2004.04.001}
}
\seealso{
Other Nucleotide distances: 
\code{\link{F81_distance}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{T92_distance}()}
}
\author{
Charles Plessy
}
\concept{Nucleotide distances}
