% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{plotApairOfChrs}
\alias{plotApairOfChrs}
\title{Plots a pair of chromosomes from a \code{GBreaks} object.}
\usage{
plotApairOfChrs(gb, chrT, chrQ = NULL)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{chrT}{A sequence name on the \emph{target} genome.}

\item{chrQ}{(Optional) A sequence name on the \emph{query} genome.}
}
\value{
Plots to the active device and returns and \code{oma::oma_layout} object
like the \code{\link[genoPlotR:plot_gene_map]{genoPlotR::plot_gene_map}} function.
}
\description{
Plots the mapping between a pair of chromosomes using the
\code{\link[genoPlotR:plot_gene_map]{genoPlotR::plot_gene_map}} function.  One sequence feature (usually a
chromosome) is selected from the \emph{target} genome.  The corresponding sequence
feature on the \emph{query} genome is either chosen automatically (being the one
with the largest fraction of the mappings) or given as a parameter.
}
\note{
In this kind of plot, the ranges from the selected sequence on the
\emph{target} genome that are not mapped to the \emph{query} sequence are not
displayed.
}
\examples{
inv       <- GRanges(c("XSR:101-180:+", "XSR:201-300:-",  "XSR:320-400:+"))
inv$query <- GRanges(c( "S1:101-200",    "S1:201-300",     "S1:301-400"))
inv <- GBreaks(inv)
plotApairOfChrs(inv, "XSR")

}
\author{
Charles Plessy
}
