% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPlotR.R
\name{plotApairOfChrs}
\alias{plotApairOfChrs}
\title{Plots a pair of chromosomes from a \code{GBreaks} object.}
\usage{
plotApairOfChrs(
  gb,
  chrT = NULL,
  chrQ = NULL,
  dna_seg_scale = TRUE,
  dna_seg_labels = NULL,
  ...
)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{chrT}{A sequence name on the \emph{target} genome.  Defaults to the first
\emph{sequence level} of the \code{gb} object.}

\item{chrQ}{(Optional) A sequence name on the \emph{query} genome.  Defaults to
the longest cumulative match on \code{chrT}.}

\item{dna_seg_scale}{Plot coordinats and a scale bar (defaults to \code{TRUE}).}

\item{dna_seg_labels}{A character vector to override default labels for
sequence names.}

\item{...}{Further arguments are passed to \code{plot_gene_map}.}
}
\value{
Plots to the active device and returns and \code{oma::oma_layout} object
like the \code{\link[genoPlotR:plot_gene_map]{genoPlotR::plot_gene_map}} function.
}
\description{
Plots the mapping between a pair of chromosomes using the
\code{\link[genoPlotR:plot_gene_map]{genoPlotR::plot_gene_map}} function.  One sequence feature (usually a
chromosome) is selected from the \emph{target} genome.  The corresponding sequence
feature on the \emph{query} genome is either chosen automatically (being the one
with the largest fraction of the mappings) or given as a parameter.
}
\note{
In this kind of plot, the ranges from the selected sequence on the
\emph{target} genome that are not mapped to the \emph{query} sequence are not
displayed.
}
\examples{
plotApairOfChrs(exampleInversion)
plotApairOfChrs(exampleDeletion)

# Labels can be overriden.
plotApairOfChrs(exampleDeletion, dna_seg_labels = c("over...", "...ridden"))

}
\seealso{
Other genoPlotR functions: 
\code{\link{gb2comp}()},
\code{\link{gb2dna_seg}()},
\code{\link{gb2xlim}()},
\code{\link{gr2dna_seg}()}

Other plot functions: 
\code{\link{makeOxfordPlots}()}

Other Structural variants: 
\code{\link{StructuralVariants}},
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flagTranslocations}()}
}
\author{
Charles Plessy
}
\concept{Structural variants}
\concept{genoPlotR functions}
\concept{plot functions}
