% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{removeDoubleInversions}
\alias{removeDoubleInversions}
\title{Remove double inversions}
\usage{
removeDoubleInversions(gb, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{detect}{Run again \code{flagDoubleInversions()} if \code{TRUE}, else reuse the \code{Dbl}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object after removing the ranges that are
doubly inverted and coalescing the resulting collinear ranges.
}
\description{
Remove double inversions
}
\examples{
removeDoubleInversions(exampleDoubleInversion1)

}
\seealso{
Other Inversion functions:
\code{\link[=filterDoubleInversions]{filterDoubleInversions()}},
\code{\link[=filterInversions]{filterInversions()}},
\code{\link[=flagDoubleInversions]{flagDoubleInversions()}},
\code{\link[=flagInversions]{flagInversions()}},
\code{\link[=flagPairs]{flagPairs()}},
\code{\link[=flagTwinInversions]{flagTwinInversions()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=leftInversionGaps]{leftInversionGaps()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=showInversions]{showInversions()}}

Other modifier functions:
\code{\link[=bridgeRegions]{bridgeRegions()}},
\code{\link[=coalesce_contigs]{coalesce_contigs()}},
\code{\link[=flipInversions]{flipInversions()}},
\code{\link[=forceSeqLengths]{forceSeqLengths()}},
\code{\link[=guessSeqLengths]{guessSeqLengths()}},
\code{\link[=keepLongestPair]{keepLongestPair()}},
\code{\link[=matchPairs]{matchPairs()}},
\code{\link[=mergeSeqLevels]{mergeSeqLevels()}},
\code{\link[=removeInversions]{removeInversions()}},
\code{\link[=removeTranslocations]{removeTranslocations()}},
\code{\link[=reverse]{reverse()}},
\code{\link[=splitSeqLevel]{splitSeqLevel()}},
\code{\link[=swap]{swap()}}
}
\concept{Inversion functions}
\concept{modifier functions}
