% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{removeInversions}
\alias{removeInversions}
\title{Remove inversions}
\usage{
removeInversions(gb, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{detect}{Run again \code{flagInversions()} if \code{TRUE}, else reuse the \code{inv}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object after removing the ranges that are
inverted and coalescing the resulting collinear ranges.
}
\description{
Remove inversions
}
\examples{
removeInversions(exampleInversion)

}
\seealso{
Other Inversion functions: 
\code{\link{filterDoubleInversions}()},
\code{\link{filterInversions}()},
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flipInversions}()},
\code{\link{leftInversionGaps}()},
\code{\link{showInversions}()}

Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{flipInversions}()},
\code{\link{forceSeqLengths}()},
\code{\link{guessSeqLengths}()},
\code{\link{keepLongestPair}()},
\code{\link{mergeSeqLevels}()},
\code{\link{removeTranslocations}()},
\code{\link{reverse}()},
\code{\link{splitSeqLevel}()},
\code{\link{swap}()}
}
\concept{Inversion functions}
\concept{modifier functions}
