% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{removeTranslocations}
\alias{removeTranslocations}
\title{Remove translocations}
\usage{
removeTranslocations(gb, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{detect}{Run again \code{flagTranslocations()} if \code{TRUE}, else reuse the \code{tra}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object after removing the ranges that are
translocated and coalescing the resulting collinear ranges.
}
\description{
Remove translocations
}
\examples{
removeTranslocations(exampleTranslocation)

}
\seealso{
Other Translocation functions: 
\code{\link{filterTranslocations}()},
\code{\link{flagTranslocations}()},
\code{\link{showTranslocations}()}

Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{flipInversions}()},
\code{\link{forceSeqLengths}()},
\code{\link{guessSeqLengths}()},
\code{\link{keepLongestPair}()},
\code{\link{mergeSeqLevels}()},
\code{\link{removeInversions}()},
\code{\link{reverse}()},
\code{\link{splitSeqLevel}()},
\code{\link{swap}()}
}
\concept{Translocation functions}
\concept{modifier functions}
