% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\alias{reverse,GRanges-method}
\alias{reverse,GBreaks-method}
\title{Reverse genomic ranges}
\usage{
reverse(x, ...)

\S4method{reverse}{GRanges}(x, ...)

\S4method{reverse}{GBreaks}(x, query = FALSE, ...)
}
\arguments{
\item{x}{A \code{GBreaks} or a \code{GRanges} object}

\item{...}{Additional arguments (ignored)}

\item{query}{On \code{GBreaks} objects, operate on othe \emph{query} or the \emph{target}.}
}
\value{
Returns the modified \code{GBreaks} or \code{GRanges} object.
}
\description{
Reverse genomic ranges by flipping the strand and moving the origin of the
coordinate system to the opposite side.
}
\examples{
gb <-       GRanges(c("chr1:101-180:+", "chr1:201-300:-",  "chr2:120-400:+"))
gb$query <- GRanges(c("chrA:101-180",   "chrB:201-300",    "chrA:120-400"))
seqlengths(gb)       <- c(500, 1000)
seqlengths(gb$query) <- c(500, 1000)
gb <- GBreaks(gb)

gb
reverse(gb)
reverse(gb, query = TRUE)
# Do not run reverse on the query directly as it will miss strand information
reverse(gb$query)

}
\seealso{
See also the \code{\link[IRanges:reverse-methods]{IRanges::reverse}} function.

Other modifier functions: 
\code{\link{coalesce_contigs}()},
\code{\link{mergeSeqLevels}()},
\code{\link{swap}()}
}
\author{
Charles Plessy
}
\concept{modifier functions}
