% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\alias{reverse,GRanges-method}
\alias{reverse,GBreaks-method}
\title{Reverse genomic ranges}
\usage{
reverse(x, ...)

\S4method{reverse}{GRanges}(x, ...)

\S4method{reverse}{GBreaks}(x, query = FALSE, ...)
}
\arguments{
\item{x}{A \code{GBreaks} or a \code{GRanges} object}

\item{...}{Additional arguments (ignored)}

\item{query}{On \code{GBreaks} objects, operate on on the \emph{query} or the \emph{target}.}
}
\value{
Returns the modified \code{GBreaks} or \code{GRanges} object.
}
\description{
Reverse genomic ranges by flipping the strand and moving the origin of the
coordinate system to the opposite side.
}
\examples{
exampleInsertion
reverse(exampleInsertion)
reverse(exampleInsertion, query = TRUE)

}
\seealso{
See also the \code{\link[IRanges:reverse-methods]{IRanges::reverse}} function.

Other modifier functions: 
\code{\link{coalesce_contigs}()},
\code{\link{guessSeqLengths}()},
\code{\link{mergeSeqLevels}()},
\code{\link{swap}()}
}
\author{
Charles Plessy
}
\concept{modifier functions}
