% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffoldByFlipAndMerge.R
\name{scaffoldByFlipAndMerge}
\alias{scaffoldByFlipAndMerge}
\title{Guided scaffolding}
\usage{
scaffoldByFlipAndMerge(gr, guide, drop = FALSE)
}
\arguments{
\item{gr}{A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object.}

\item{guide}{A named list of data frames of contig names and orientations.}

\item{drop}{Drop the contigs that were not included in the guide.}
}
\description{
Guided scaffolding
}
\examples{
(gr <- exampleTranslocation |> swap())
(g <- list(chrBC = data.frame(contig = c("chrB", "chrC"), orientation = c(1,-1))))
scaffoldByFlipAndMerge(g, gr)

}
\concept{family scaffolding functions}
