% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagInversions.R
\name{showInversions}
\alias{showInversions}
\title{Show inversions and their flanking blocks.}
\usage{
showInversions(gb, rename = TRUE, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}}.}

\item{rename}{Replace range names by their numeric order before subsetting.}

\item{detect}{Run again \code{flagInversions()} if \code{TRUE}, else reuse the \code{inv}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
part of an inversion triplet have been discarded.  If the object was missing
the \code{inv} metadata column, return the object after discarding all of its
ranges.
}
\description{
Show inversions and their flanking blocks.
}
\examples{
showInversions(exampleInversion)

}
\seealso{
Other Inversion functions: 
\code{\link{filterDoubleInversions}()},
\code{\link{filterInversions}()},
\code{\link{flagDoubleInversions}()},
\code{\link{flagInversions}()},
\code{\link{flagPairs}()},
\code{\link{flipInversions}()},
\code{\link{leftInversionGaps}()}
}
\concept{Inversion functions}
