% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translocations.R
\name{showTranslocations}
\alias{showTranslocations}
\title{Show translocations and their flanking blocks.}
\usage{
showTranslocations(gb, rename = TRUE, detect = TRUE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{rename}{Replace range names by their numeric order before subsetting.}

\item{detect}{Run again \code{flagTranslocations()} if \code{TRUE}, else reuse the \code{tra}
flag or fail if it is absent.}
}
\value{
Returns the \code{GBreaks} object in which all ranges that are not
part of a translocation triplet have been discarded.  If the object was missing
the \code{tra} metadata column, return the object after discarding all of its
ranges.
}
\description{
Show translocations and their flanking blocks.
}
\examples{
showTranslocations(exampleTranslocation)

}
\seealso{
Other Translocation functions: 
\code{\link{filterTranslocations}()},
\code{\link{flagTranslocations}()},
\code{\link{removeTranslocations}()}
}
\concept{Translocation functions}
