% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Sliding window}
\usage{
slidingWindow(
  gb,
  factor = 100,
  windowSize = NULL,
  stepSize = NULL,
  merged = FALSE,
  cut = TRUE
)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{windowSize}{The size of the window.}

\item{stepSize}{The size of the step.}

\item{merged}{set TRUE to output a \link{'GBreaks'} object}

\item{cut}{whether cut the alignments out of the window}

\item{type}{slide the window on the reference genome or query genome (reference/query)}
}
\value{
Returns a list of \code{GBreaks} objects, one for each window.
}
\description{
This function is used to subset a \code{GBreaks} object into a list of \code{GBreaks}
using a sliding window.
}
\examples{
exdata_Sac <- system.file("extdata/SacCer3__SacPar.gff3.gz", package = "GenomicBreaks")
gb <- load_genomic_breaks(exdata_Sac, BSgenome.Scerevisiae.UCSC.sacCer3::Scerevisiae)
slided_gb <- slidingWindow_by_bp(gb, windowSize = 1e5, stepSize = 5e4)
slided_gb[[1]]

}
