% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_randomisation_index.R
\name{strand_randomisation_index}
\alias{strand_randomisation_index}
\title{Index measuring how 'random' the alignment strand has become.}
\usage{
strand_randomisation_index(gb, tiles = NULL)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{tiles}{A number of tiles}
}
\value{
Returns a numeric value between 0 and 1.
}
\description{
In groups of species where major changes of gene order happened but genes
tended to stay on the same chromosome, a marked feature is that the strand
on which homologous sequences align appears to be randomised.  This index
expresses it with a number.
}
\details{
For each feature in the \emph{target} genome, the total length of alignments on
the plus strand is subtracted from that on the minus strand, and the absolute
value is taken (because there is no guarantee that homologous chromosomes are
sequenced in the same direction in different assemblies).  This value is
normalized by the total number of aligned bases.  A weighted mean is then
computed across all features, using each feature’s total aligned bases as its
weight.  Thus, a number close to 1 is expected for closely related genomes.
}
\note{
This index is designed for comparison of chromosomal assemblies that
have a strong conservation of synteny in the sense most homologous genes
are mapped on pairs of homologous chromosomes.  In 2025, a new \code{tiles} option
was added to make the index more robust to assembly errors and karyotype
changes.
}
\examples{
strand_randomisation_index(exampleInversion)

}
\references{
Plessy C, Mansfield MJ, Bliznina A, Masunaga A, West C, Tan Y,
Liu AW, Grašič J, Del Río Pisula MS, Sánchez-Serna G, Fabrega-Torrus M,
Ferrández-Roldán A, Roncalli V, Navratilova P, Thompson EM, Onuma T, Nishida
H, Cañestro C, Luscombe NM (2024). Extreme genome scrambling in marine
planktonic \emph{Oikopleura dioica} cryptic species. \emph{Genome Research}, 34(3),
426–440. \url{https://doi.org/10.1101/gr.278295.123}. PMID: 38621828
}
\seealso{
Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=synteny_index]{synteny_index()}},
\code{\link[=tau_index]{tau_index()}}
}
\author{
Charles Plessy
}
\concept{Similarity indexes}
