% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_randomisation_index.R
\name{strand_randomisation_index}
\alias{strand_randomisation_index}
\title{Index measuring how 'random' the alignment strand has become.}
\usage{
strand_randomisation_index(gb, tiles = NULL)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{tiles}{A number of tiles}
}
\value{
Returns a numeric value between 0 and 1.
}
\description{
In groups of species where major changes of gene order happened but genes
tended to stay on the same chromosome, a marked feature is that the strand
on which homologous sequences align appears to be randomised.  This index
expresses it with a number.
}
\details{
For each feature of the \emph{target} genome, the total length of its alignments
on the plus strand is subtracted from the total length of its alignments
on the minus strand.  The sign of this number is removed because there is
no guarantee that homologous chromosomes are sequenced in the same direction
in different assemblies.  To make features comparable, this number is then
normalised by total number of aligned bases.  Then, the computed numbers for
all the \emph{target}'s features are averaged with a weighted mean, where the
weight is the total number of aligned bases of that feature.  Thus, a number
close to 1 is expected for closely related genomes.
}
\note{
This index is designed for comparison of chromosomal assemblies that
have a strong conservation of synteny in the sense most homologous genes
are mapped on pairs of homologous chromosomes.
}
\examples{
strand_randomisation_index(exampleInversion)

}
\seealso{
Other Similarity indexes: 
\code{\link{F81_distance}()},
\code{\link{GOC}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{T92_distance}()},
\code{\link{correlation_index}()},
\code{\link{karyotype_index}()},
\code{\link{slidingWindow}()},
\code{\link{synteny_index}()}
}
\concept{Similarity indexes}
