% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandNames.R
\name{strandNames}
\alias{strandNames}
\title{Sequence name with strand information}
\usage{
strandNames(gb, flip = FALSE, query = FALSE)
}
\arguments{
\item{gb}{A \code{GBreaks} object or a \code{\link{GRangesList}} of \code{GBreaks}
objects.}

\item{flip}{Flip the strand names.}

\item{query}{Work on the query genome.}
}
\value{
Returns a character vector, or a list of character vectors if the
input was a \code{\link{GRangesList}}.
}
\description{
Extract sequence names and pastes strand information to it.
}
\examples{
strandNames(exampleColinear)
strandNames(exampleColinear, query = TRUE)

}
\seealso{
Other scaffolding functions: 
\code{\link{longestMatchesInTarget}()},
\code{\link{mergeSeqLevels}()},
\code{\link{scaffoldByFlipAndMerge}()}
}
\author{
Charles Plessy
}
\concept{scaffolding functions}
