% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralVariants.R
\name{StructuralVariants}
\alias{StructuralVariants}
\alias{exampleColinear}
\alias{exampleNotColinear}
\alias{exampleColinear3}
\alias{exampleInversion}
\alias{exampleTranslocation}
\alias{exampleInsertion}
\alias{exampleDeletion}
\title{Structural Variants}
\format{
\code{\link{GBreaks}} objects described in more details in the
\emph{Structural Variants} vignette, available with the command
\code{vignette("Structural Variants", package = "GenomicBreaks")}.
}
\description{
Concept examples of structural variants for test and
demonstration purpose
}
\details{
The variants represented here are the ones that can be represented
in a \code{\link{GenomicBreaks}} object.  Short ones that are entirely contained in an
aligned region, for instance as a gap, are not covered
}
\examples{
exampleColinear
exampleNotColinear
exampleColinear3
exampleInversion
exampleTranslocation

}
\seealso{
Other Structural variants: 
\code{\link{flagInversions}()},
\code{\link{plotApairOfChrs}()}
}
\author{
Charles Plessy
}
\concept{Structural variants}
