% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap target and query genomes}
\usage{
swap(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object}
}
\value{
A new \code{GBreaks} object representing the input's \emph{query} ranges,
with the input's \emph{target} ranges in the \emph{query} slot.  The strand information
is unchanged.
}
\description{
Produce a new object in which the information about the \emph{target} and
\emph{query} genomes have been inverted.
}
\examples{
swap(exampleColinear3)

}
\seealso{
Other modifier functions: 
\code{\link{coalesce_contigs}()},
\code{\link{guessSeqLengths}()},
\code{\link{mergeSeqLevels}()},
\code{\link{reverse}()}
}
\concept{modifier functions}
