% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap reference and query genomes}
\usage{
swap(gr_ob)
}
\arguments{
\item{gr_ob}{Genomic ranges representing a pairwise genome alignment.}
}
\value{
Returns a new GRanges object representing the input's query ranges
with the input's reference ranges in the query slot.  The strand information
is transferred so that the query ranges are strandless and the reference
are stranded.
}
\description{
Produce a new object in which the information about the reference and
query genomes have been inversed.
}
\examples{
gr       <- GenomicRanges::GRanges(c("chr1:100-200:+", "chr1:300-400:+", "chr1:500-600:-"))
gr$query <- GenomicRanges::GRanges(c("chr5:300-400",   "chr6:100-200",   "chr7:200-300"))

}
