% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap target and query genomes}
\usage{
swap(gb, sort = FALSE)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}

\item{sort}{Sort the returned object, ignoring strand.}
}
\value{
A new \code{GBreaks} object representing the input's \emph{query} ranges,
with the input's \emph{target} ranges in the \emph{query} slot.  The strand information
is unchanged.
}
\description{
Produce a new object in which the information about the \emph{target} and
\emph{query} genomes have been inverted.  As a large number of \emph{GenomicBreaks}
functions require \code{GBreaks} objects that were sorted with strand information
ignored, a \code{sort} option is provided to do that.
}
\examples{
swap(exampleDeletion)
swap(exampleDeletion, sort = TRUE)

}
\seealso{
Other modifier functions: 
\code{\link{bridgeRegions}()},
\code{\link{coalesce_contigs}()},
\code{\link{flipInversions}()},
\code{\link{forceSeqLengths}()},
\code{\link{guessSeqLengths}()},
\code{\link{keepLongestPair}()},
\code{\link{mergeSeqLevels}()},
\code{\link{removeDoubleInversions}()},
\code{\link{removeInversions}()},
\code{\link{removeTranslocations}()},
\code{\link{reverse}()},
\code{\link{splitSeqLevel}()}
}
\concept{modifier functions}
