% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T92_distance.R
\name{T92_distance}
\alias{T92_distance}
\title{Tamura 1992 distance}
\usage{
T92_distance(m)
}
\arguments{
\item{m}{A matrix of counts or probabilities for bases of the \emph{target} genome
to be aligned to bases on the \emph{query} genome.  As a convenience it can also
receive a list produced by the \code{\link[=readTrainFile]{readTrainFile()}} function, containing this
matrix.}
}
\value{
Returns a numeric value show the evolutionary distance between two genomes. the larger the value, the more different the two genomes are.
}
\description{
The Tamura 1992 (T92) distance extends the K80 distance by taking \code{GC} content into account.  It is calculated as \eqn{-h \ln \left(1 - \frac{p}{h} - q\right) - \frac{1}{2} \times (1 - h) \ln\left(1 - 2 q\right)}, where \eqn{p} is the probability of transition, \eqn{q} the probability of transversion, \eqn{h = 2\theta (1 - \theta)} and \eqn{\theta} is the \code{GC} content. See the \href{https://en.wikipedia.org/wiki/Models_of_DNA_evolution#T92_model_(Tamura_1992)}{Wikipedia} for more details.
}
\examples{
parameters <- readTrainFile(system.file("extdata/example.train", package = "GenomicBreaks"))
T92_distance(parameters)

}
\references{
Tamura, K. (1992). "Estimation of the number of nucleotide substitutions when there are strong transition-transversion and G+C-content biases." \emph{Molecular Biology and Evolution}, 9(4), 678–687. DOI: \href{https://doi.org/10.1093/oxfordjournals.molbev.a040752}{10.1093/oxfordjournals.molbev.a040752}
}
\seealso{
Other Nucleotide distances: 
\code{\link{F81_distance}()},
\code{\link{JC69_distance}()},
\code{\link{K80_distance}()},
\code{\link{P_distance}()},
\code{\link{gapProportion}()}
}
\author{
Zikun Yang
}
\concept{Nucleotide distances}
