% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_index.R
\name{tau_index}
\alias{tau_index}
\title{Tau index}
\usage{
tau_index(gb)
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object.}
}
\value{
Returns a numeric value between 0 and 1.
}
\description{
Tissue-specificity index from Yanai and coll., 2005 repurposed in 2022 by
Ranz and coll. for synteny analysis.
}
\details{
In 2022, Ranz and coll., the tau index was computed for each chromosome in a
\emph{target} genome compared to a \emph{query}, with a formula such as: for a given
\emph{target} feature, count its one-to-one orthologues on every feature \eqn{i}
of the \emph{query} genome, normalize these counts by dividing by the largest
count, subtract each normalized value from one, sum the results, and divide
by the number \eqn{n} of \emph{query} features minus one.

\deqn{\frac{1}{n-1}\sum_{i=1}^{n}\left(1 - \frac{x_i}{\max(x)}\right)}{sum(1 - x / max(x)) / (length(x) - 1)}

Here, the index is extended to the whole genome comparisons by computing it
for each feature of the \emph{target} genome and returning the average weighted
by feature length.
}
\note{
Note that calculating the tau index on whole-genome nucleotide
alignments is not expected to produce meaningful results.  This function is
more useful when comparing the positions of protein orthologues.
}
\examples{
tau_index(exampleTranslocation)
GenomicBreaks:::.tau_index(c(0,8,0,0,0,2,0,2,0,0,0))

}
\references{
Yanai I, Benjamin H, Shmoish M, Chalifa-Caspi V, Shklar M, Ophir
R, Bar-Even A, Horn-Saban S, Safran M, Domany E, Lancet D, Shmueli O. (2005).
Genome-wide midrange transcription profiles reveal expression level
relationships in human tissue specification. \emph{Bioinformatics}, 21(5):650-659.
\doi{10.1093/bioinformatics/bti042}. PMID: 15388519

Ranz JM, González PM, Su RN, Bedford SJ, Abreu-Goodger C, Markow
T. (2022).  Multiscale analysis of the randomization limits of the
chromosomal gene organization between Lepidoptera and Diptera.
\emph{Proc Biol Sci}, 289(1967):20212183. \doi{10.1098/rspb.2021.2183}.
PMID: 35042416
}
\seealso{
Other Similarity indexes:
\code{\link[=F81_distance]{F81_distance()}},
\code{\link[=GOC]{GOC()}},
\code{\link[=HKY85_distance]{HKY85_distance()}},
\code{\link[=JC69_distance]{JC69_distance()}},
\code{\link[=JC69_distance_allseq]{JC69_distance_allseq()}},
\code{\link[=K80_distance]{K80_distance()}},
\code{\link[=K80_gap_distance]{K80_gap_distance()}},
\code{\link[=P_distance]{P_distance()}},
\code{\link[=T92_distance]{T92_distance()}},
\code{\link[=TN93_distance]{TN93_distance()}},
\code{\link[=breakpointGraphProperties]{breakpointGraphProperties()}},
\code{\link[=correlation_index]{correlation_index()}},
\code{\link[=inversionDistance]{inversionDistance()}},
\code{\link[=inversionEstimate_BD]{inversionEstimate_BD()}},
\code{\link[=karyotype_index]{karyotype_index()}},
\code{\link[=logDet_distance]{logDet_distance()}},
\code{\link[=slidingWindow]{slidingWindow()}},
\code{\link[=strand_randomisation_index]{strand_randomisation_index()}},
\code{\link[=synteny_index]{synteny_index()}}
}
\author{
Charles Plessy
}
\concept{Similarity indexes}
