% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wholeGenomeClassification.R
\name{wholeGenomeClassification}
\alias{wholeGenomeClassification}
\title{Whole-genome classification object}
\usage{
wholeGenomeClassification(gb, coa = coalesce_contigs(gb))
}
\arguments{
\item{gb}{A \code{\link{GBreaks}} object representing a one-to-one whole genome
alignment.}

\item{coa}{The coalesced one-to-one alignment.  If not provided, it will be\
computed on-the-fly with the \code{\link{coalesce_contigs}} function.}
}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object representing the \emph{target}
genome, inheriting its \emph{sequence information} (\code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo}}).
The class of each region is indicated by a factor in the \code{type} metadata
column.
}
\description{
Classifies aligned genomic regions as \emph{isolated} or \emph{collinear}, and
unaligned regions as \emph{breakpoint} or \emph{bridge} regions.  The \emph{bridge} regions
are always flanked by \emph{collinear alignments} and the \emph{isolated alignments}
are always flanked by \emph{breakpoint regions}.
}
\examples{
exampleColinear5 |> wholeGenomeClassification()

}
\seealso{
Other Reducing functions: 
\code{\link{cleanGaps}()},
\code{\link{get_bps}()}
}
\author{
Charles Plessy

Michael Mansfield
}
\concept{Reducing functions}
