% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipWithNext.R
\name{zipWithNext}
\alias{zipWithNext}
\title{Zip elements with the next ones}
\usage{
zipWithNext(x, step = 1)
}
\arguments{
\item{x}{Object with a vector structure.}

\item{step}{Remove the first \code{step} entries from the object before zipping
it with itself.}
}
\value{
A \code{\link[S4Vectors:Pairs-class]{S4Vectors::Pairs}} object with a copy of \code{x} in the \code{first} slot,
and \code{x} with \code{step} elements removed from its head and complemented with
\code{NAs} in the \code{second} slot.
}
\description{
Zip an object with itself after inserting a variable number of steps.  For
instance, with a step of 1, \emph{a} \emph{b} \emph{c} would become \emph{a,b}, \emph{b,c}, \emph{c,NA}.
}
\details{
This is a programming pattern that I use a lot in this package, to prepare
vectors fro parallel processing.  I typically use it on metadata columns
of \code{GBreaks} objects.  I created this function because I always forget if I
should use \code{head} or \code{tail}, which causes bugs or delays in development.
}
\examples{
zipWithNext(LETTERS, step = 2)

}
